### 2025-01-17 #7679 @arfon
@feihelab – thanks for your submission to JOSS. On first inspection, it's not clear if this submission meets our requirements around testing and packaging. I'm going to flag this for a scope review, this will likely take a few weeks to complete.
