### 2025-03-06 #7679 @raj-gun
@arfon The bot has generated an error for generating the PDF. The JOSS github actions can generate the pdf of the draft successfully. I don't know what the error means. The error message is given below.

Run gh issue comment 7679 --body ":warning: [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/13697519256?check_suite_focus=true). "
https://github.com/openjournals/joss-reviews/issues/767[9](https://github.com/openjournals/joss-papers/actions/runs/13697519256/job/38303235669#step:6:10)#issuecomment-2703562792
