### 2025-03-14 #7679 @boisgera
> I need a bit of time to think about how to shorten the paper, sorry for keeping you all waiting, I'm in the middle of another paper. I hope that is okay. I will get back to you as soon as possible.

Sure, take the time you need. We'll start the review formally when you are ready (I need to find at least a second reviewer to start the review anyway).
