### 2025-04-08 #7679 @boisgera
Hi @raj-gun,

Do you want me to start formally the review? @JAuriol81 and @brivadis are ready if you are.

The paper still needs ultimately to be shortened to fit into the ~1000-2000 words range, but you can do it during the review if that's ok with you.

Best regards,

Sébastien
