### 2025-04-08 #7679 @raj-gun
Hi Sébastien,

Sorry for the delay in shortening the paper and the inconvenience. Yes if it is okay with you, we can start the formal review.

Thank you
Kind regards
Raj
