### 2025-04-08 #8028 @editorialbot
[REVIEW]: NonSysId: Nonlinear System Identification with Improved Model Term Selection for NARMAX Models
**Submitting author:** @feihelab (<a href="http://orcid.org/0000-0001-9176-6674">Fei He</a>)
**Repository:** https://github.com/raj-gun/NonSysID
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
