### 2025-04-11 #8028 @brivadis
Hello ! In the review checklist there is **Community guidelines**.
There does not seem to be clear guidelines for third parties in the repository.
