### 2025-04-11 #8028 @brivadis
**Automated tests:** Do you plan to add automated unitary tests to verify some core functions of the software ?
