### 2025-04-11 #8028 @brivadis
**Installation** will be checked once this issue is solved:
- [https://github.com/raj-gun/NonSysID/issues/6](https://github.com/raj-gun/NonSysID/issues/6)
