### 2025-04-11 raj-gun/NonSysID#10 @JAuriol81
Documentation for the arguments of the function "NonSysID"
I believe a documentation file would be important to clearly explain the meaning of the different arguments of the function
[model, Mod_Val_dat, iOFR_table_lin, iOFR_table_nl, best_mod_ind_lin, best_mod_ind_nl, val_stats] = ...
    NonSysID(mod_type,u_ID,y_ID,na1,na2,nb1,nb2,nl_ord_max,is_bias,n_inpts,KSA_h,RCT,x_iOFR,stp_cri,D1_thresh,displ,sim,parall);
