### 2025-04-11 #8028 @boisgera
> **Substantial scholarly effort:** The repository is quite recent, and has been pushed by 1 contributor in about 2 months. I let [@boisgera](https://github.com/boisgera) decide about that point.

The `cloc` commands finds 1783 lines of Matlab code currently, which is in the small range, does not raises any concern (< 1000 locs usually triggers a deeper review).

About the project age: the command `git log --reverse --pretty='format:%H %ad' --date=short --numstat -- *.m` shows me that an initial commit with >3k locs of Matlab files which strongly suggests that the project was started a while before the git repo was initialized (see also : https://github.com/raj-gun/NonSysID/commit/996059c12ed763b7caf27120779f46489fa87406).

@feihelab Can you confirm this?

Anyway, from this (arguably superficial) analysis, I don't see the scholarly effort as a problem here.
