### 2025-05-12 #8028 @boisgera
Hi @feihelab! 👋

Could you address [the issues opened by @JAuriol81 and @brivadis](https://github.com/raj-gun/NonSysID/issues?q=is%3Aissue%20state%3Aopen%20(author%3AJAuriol81%20OR%20author%3Abrivadis)) so that we can move the review forward?

Thank you in advance.

Sébastien (@boisgera)
