### 2025-05-14 #8028 @raj-gun
Hi everyone,

> Hi [@feihelab](https://github.com/feihelab)! 👋
>
> Could you address [the issues opened by @JAuriol81 and @brivadis](https://github.com/raj-gun/NonSysID/issues?q=is%3Aissue%20state%3Aopen%20(author%3AJAuriol81%20OR%20author%3Abrivadis)) so that we can move the review forward?
>
> Thank you in advance.
>
> Sébastien ([@boisgera](https://github.com/boisgera))

> **Installation** will be checked once this issue is solved:
>
> * [addpath issue raj-gun/NonSysID#6](https://github.com/raj-gun/NonSysID/issues/6)

My sincere apologies @boisgera and @brivadis, for the extremely late reply.

The line "addpath('\...\NonSysID');" is there to add the path of the folder containing the NonSysID package.
For example, "addpath('C:\Users\user_name\Documents\NonSysID')".

I'm sorry that this wasn't mentioned.

Thank you
Kind regards
Raj
