### 2025-07-09 1c51ca7
Update paper.md
