+++ paper/paper.md
@@ -33,49 +33,17 @@ System identification involves constructing mathematical models of dynamic syste
 System identification is a field at the intersection of control theory, dynamic systems theory and machine learning that seeks to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. 
 Generally, system identification has two primary objectives [@ljung1998system;@billings2013a], (i) to accurately map the system's inputs and outputs, allowing for the prediction of new, unseen data, and (ii) to capture the underlying dynamics of the system within the model.
 
-The dynamic models generated through system identification can be either discrete or continuous time models [@UNBEHAUEN1997]. This paper centres on widely-used discrete-time models, specifically nonlinear auto-regressive models with exogenous inputs (NARX), where the ARX model is a linear variant of the NARX framework. These input-output time-series models predict future outputs of a system based on its historical input and output instances. NARX models have been applied extensively to model and analyse complex systems in fields such as control, condition monitoring, advanced manufacturing and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Moreover, it has been demonstrated that the NARX model has equivalence to a recurrent neural network (RNN) [@Sum1999]. Extending the NARX model to incorporate a noise model, we obtain the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model.  
+The dynamic models generated through system identification can be either discrete or continuous time models [@UNBEHAUEN1997]. This paper centres on widely-used discrete-time models, specifically nonlinear auto-regressive models with exogenous inputs (NARX), where the ARX model is a linear variant of the NARX framework. These input-output time-series models predict future outputs of a system based on its historical input and output instances. NARX models have been applied extensively to model and analyse complex systems in fields such as control, condition monitoring, advanced manufacturing and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Moreover, it has been demonstrated that the NARX model is equivalent to a recurrent neural network (RNN) [@Sum1999]. Extending the NARX model to incorporate a noise model, we obtain the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model.  
 
-Recently two open-sourced packages have been introduced, `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive. However, they are based on the original forward regression orthogonal least squares (OFR) algorithm, which has been noted to have several limitations, as discussed in [@Piroddi2003;@Mao1997]. These concerns primarily involve over-fitting and inaccurate long-horizon predictions, particularly when the input fails to sufficiently excite the actual system. Additionally, in some applications, acquiring extra data for cross-validation may be infeasible. As a result, developing parsimonious models that can generalise well to unseen data becomes crucial in such cases. This paper introduces the `NonSysId` package, which incorporates an enhanced model selection process to address these challenges.
[…]
