### 2025-07-09 b03146c
Update paper.md
