+++ paper/paper.md
@@ -53,7 +53,7 @@ In this example, $\phi_{1}(t)=y(t-1)$ and $\phi_{2}(t)=u(t-2)$ have a degree of 
         \Big\} ,
 \end{multline}
 
-from which a NARX model structure, such as that in \autoref{eq:narx_exmpl}, can be identified. Once the model structure is identified, the next step is to estimate the model parameters. However, determining the appropriate linear and nonlinear terms to include in the model structure is critical to achieving parsimonious models. This is particularly important in the nonlinear cases (Chapter 1, [@billings2013a]), as the inclusion of  unnecessary model terms can result in a model that erroneously captures dynamics that do not belong to the underlying system [@AGUIRRE1995;@mendes1998a].
+from which a NARX model structure, such as that in \autoref{eq:narx_exmpl}, can be identified. However, determining the appropriate linear and nonlinear terms to include in the model structure is critical to achieving parsimonious models. This is particularly important in the nonlinear cases (Chapter 1, [@billings2013a]), as the inclusion of  unnecessary model terms can result in a model that erroneously captures dynamics that do not belong to the underlying system [@AGUIRRE1995;@mendes1998a].
 
 The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regression OLS (FROLS) [@chen1989b;@billings1987a], is based on the Orthogonal Least Squares (OLS). When combined with an appropriate term selection criterion [@korenberg1988a;@WANG1996;@hong2003], it efficiently selects model terms (regressors) in a forward selection manner (sequentially according to the term selection criterion), facilitating the development of a parsimonious model. The OFR/FROLS algorithm evaluates the impact of each term on the output independently of the influence of other terms, which is achieved through  orthogonalization procedures. This evaluation relies on the chosen term selection criterion, allowing for the sequential inclusion of appropriate terms in the final model using a forward selection approach. The most commonly used and widely accepted model term selection criterion used in the OFR algorithm is the error reduction ratio (ERR). During the forward selection procedures, the ERR selects the term that maximises explained variance, thereby  maximising the goodness of fit. Over the years, many variants of the OFR have been proposed. However, concerns persist regarding the original OFR algorithm (OFR-ERR), which relies on the ERR for term selection [@Piroddi2003;@Mao1997], for example,
 
@@ -97,7 +97,7 @@ In step 2, each model $\Tilde{m}_i \in \Tilde{\mathcal{M}}$, $i=1,\dots,p$, is t
 
 ## PRESS-statistic-based term selection {#PRESS}
 
-The model must generalise effectively to unseen data during training, striking a balancing bias and variance. This can be achieved using robust cross-validation strategies. Ideally, an algorithm should optimise model generalisation without relying on a separate validation dataset. A PRESS-statistic-based [@Allen1974] term selection criterion with leave-one-out cross-validation was introduced into the OFR framework in [@WANG1996;@hong2003]. Leveraging the OLS method in OFR, the computation of the leave-one-out cross-validation errors is highly efficient [@WANG1996;@hong2003]. Integrating the PRESS-statistic-based criterion into the OFR algorithm enables the selection of regressors (model terms) that incrementally minimise the one-step ahead leave-one-out cross-validation error in a forward selection manner, effectively reducing overfitting to noise. This approach fully automates the model evaluation process, eliminating the need for additional validation data. Consequently, applying the PRESS-statistic-based term selection criterion within the $\text{iOFR}_{S}$ algorithm enhances the selection of more robust terms and improves the generalisation capabilities of the resulting models.
[…]
