### 2025-07-09 109cd27
Update paper.md
