+++ paper/paper.md
@@ -113,11 +113,11 @@ The `NonSysId` package incorporates both aforementioned methods to reduce the co
 
 - **RCT Method 2**: This method, as proposed in [@guo2015a], identifies an appropriate initial set of pre-select terms, $\mathcal{P}$, for the first iteration of the iOFR algorithm. By ensuring that $\mathcal{P} \subset \mathcal{D}''$ (containing fewer redundant terms compared to $\mathcal{P} \subseteq \mathcal{D}$), the first iteration of iOFR/$\text{iOFR}_{S}$ involves fewer orthogonalization paths originating from redundant terms. This accelerates convergence towards an optimal model [@guo2015a].
 
-![Pseudo code summarising the model identification algorithm implementing $\text{iOFR}_{S}$ in `NonSysId`. \label{fig:alg}](Figures/alg.svg){width="100%"}
-
 - **RCT Method 3**: This method combines RCT methods 1 and 2, such that $\mathcal{P} \subset \mathcal{D}''_{R}$ and $\text{iOFR}_{S}$ searches through an appropriately reduced space defined by $\mathcal{D}''_{R}$. As a result, this approach enables faster convergence of $\text{iOFR}_{S}$ to an optimal model compared to any other RCT method.
 
 - **RCT Method 4**: This method combines RCT methods 1 and 2, such that $\mathcal{P} \subset \mathcal{D}''_{R}$. However, $\text{iOFR}_{S}$ searches through the full space $\mathcal{D}''$ instead of $\mathcal{D}''_{R}$. Therefore, this technique converges the $\text{iOFR}_{S}$ faster to an optimal model compared to RCT method 2.
+
+![Pseudo code summarising the model identification algorithm implementing $\text{iOFR}_{S}$ in `NonSysId`. \label{fig:alg}](Figures/alg.svg){width="100%"}
 
 The RCT methods aim to accelerate the convergence of $\text{iOFR}_{S}$ and reduce the time required to obtain a model. Using $\mathcal{D}''_{R}$ reduces the computational time for the OFR algorithm within $\text{iOFR}_{S}$, by shortening the time needed to follow a given orthogonalization path. Additionally, fewer redundant terms in $\mathcal{P}$ lead to faster convergence of $\text{iOFR}_{S}$ and contribute to reducing time by minimizing the number of orthogonalization paths [@guo2015a]. Therefore, the most effective RCT method is 3, followed by methods 1,4 and 2. However, when reducing the search space (determining $\mathcal{D}''_{R}$), RCT methods 1 and 3 may miss some correct terms, potentially resulting in convergence to a sub-optimal model. This outcome depends on the level of white and coloured noise in the input-output data, as well as the complexity of the original system. It should be noted that RCT methods introduce additional procedures. Therefore, if $\mathcal{D}''$ is small enough, running $\text{iOFR}_{S}$ without any RCT methods may be faster. The figure below summarises Algorithm \autoref{fig:alg} in a flowchart. The following section will provide examples from the `NonSysId` package. 
 
