### 2025-07-09 a578ed2
Update paper.md
