+++ paper/paper.md
@@ -97,7 +97,7 @@ In step 2, each model $\Tilde{m}_i \in \Tilde{\mathcal{M}}$, $i=1,\dots,p$, is t
 
 ## PRESS-statistic-based term selection {#PRESS}
 
-The model must generalise effectively to unseen data during training, striking a balance between bias and variance. This can be achieved using robust cross-validation strategies. Ideally, an algorithm should optimise model generalisation without relying on a separate validation dataset. A PRESS-statistic-based [@Allen1974] term selection criterion with leave-one-out cross-validation was introduced into the OFR framework in [@WANG1996;@hong2003]. Leveraging the OLS method in OFR, the computation of the leave-one-out cross-validation errors is highly efficient [@WANG1996;@hong2003]. Integrating the PRESS-statistic-based criterion into the OFR algorithm enables the selection of regressors (model terms) that incrementally minimise the one-step ahead leave-one-out cross-validation error in a forward selection manner, effectively reducing overfitting to noise. This approach fully automates the model evaluation process, eliminating the need for additional validation data. Consequently, applying the PRESS-statistic-based term selection criterion within the $\text{iOFR}_{S}$ algorithm enhances the selection of more robust terms and improves the generalisation capabilities of the resulting models.
+The model must generalise effectively to unseen data during training, striking a balance between bias and variance. A PRESS-statistic-based [@Allen1974] term selection criterion with leave-one-out cross-validation was introduced into the OFR framework in [@WANG1996;@hong2003]. Leveraging the OLS method in OFR, the computation of the leave-one-out cross-validation errors is highly efficient [@WANG1996;@hong2003]. Integrating the PRESS-statistic-based criterion into the OFR algorithm enables the selection of regressors (model terms) that incrementally minimise the one-step ahead leave-one-out cross-validation error in a forward selection manner, effectively reducing overfitting to noise. This approach fully automates the model evaluation process, eliminating the need for additional validation data. Consequently, applying the PRESS-statistic-based term selection criterion within the $\text{iOFR}_{S}$ algorithm enhances the selection of more robust terms and improves the generalisation capabilities of the resulting models.
 
 ## $\text{iOFR}_{S}$ with reduced computational time
 
