### 2025-07-09 ffdfc28
Update paper.md
