+++ paper/paper.md
@@ -77,11 +77,9 @@ A notable feature of `NonSysId` is the implementation of the iterative-OFR (iFRO
 
 To address the concerns associated with the original OFR, the iterative-OFR (iOFR) algorithm was introduced in [@guo2015a]. To the best of our knowledge, no open-source software implementing this variant currently exists. In the original OFR, the term selection is heavily influenced by the order of orthogonalization, which can often result in incorrect terms being selected in the early stages [@guo2015a;@Mao1997]. Additionally, the order in which terms are selected in the OFR determines the orthogonalization path, resulting in a tree structure of possible models [@guo2015a;@Mao1997]. Finding a globally optimal solution would require an exhaustive search through all orthogonalization paths - an infeasible task given the factorial  growth in paths ($k!$ paths for $k$ terms). The iOFR algorithm addresses this limitation by iteratively exploring multiple orthogonalization paths and re-selecting terms to approximate a globally optimal model without exhaustive search [@guo2015a]. This approach enables the recovery of correct terms that might have been overlooked in earlier iterations. As a result, the iOFR generates several candidate models for consideration. 
 
-The iOFR procedures [@guo2015a] can be summarised as follows. Given an output vector $\mathbf{Y}$, a set of candidate terms $\mathcal{D}$ and a set of pre-select terms $\mathcal{P} \subseteq \mathcal{D}$, where $\mathcal{P} = \{ \phi_1 , \dots, \phi_p  \}$, 1) pre-select each term given in $\mathcal{P}$ as the first model term; 2) use OFR to search through $p$ orthogonalization paths resulting in a set of $p$ candidate models $\mathcal{M} = \{ m_1, \dots, m_p \}$; 3) from $\mathcal{M}$, select the best model $\overline{m}$ based on the one-step-ahead prediction error; and 4) update the set of pre-select terms $\mathcal{P}$ with the terms in $\overline{m}$. The process is repeated iteratively with the updated $\mathcal{P}$ to search through different orthogonalization paths.    
+As demonstrated in [@guo2015a], the iOFR can iteratively produce more globally optimal model structures. This is because optimal solutions are only found along orthogonalization paths that begin with a correct term [@guo2015a;@Mao1997] (candidate terms essential for accurately reconstructing dynamics of the original system [@mendes1998a;@AGUIRRE1994]). Although the best model $\overline{m}$ obtained in each iteration may be sub-optimal, it will include certain correct terms [@guo2015a]. Consequently, in subsequent iterations, $\mathcal{P}$ will contain fewer redundant model terms. This refinement ensures that, in the next iteration, a relatively greater proportion of the orthogonalization paths explored by the OFR will start from better initial terms, leading to a more robust set of models $\mathcal{M}$ [@guo2015a]. For the first iOFR iteration, the pre-select terms, $\mathcal{P}$, can be set to $\mathcal{P} = \mathcal{D}$. Since it is sufficient to focus on orthogonalization paths that begin with correct terms [@guo2015a], methods for selecting the initial set $\mathcal{P}$ will be discussed in later sections. This will make the iOFR converge faster towards an optimum while improving computational efficiency.
 
[…]
