### 2025-07-10 06f8fd9
Update paper.md
