+++ paper/paper.md
@@ -73,7 +73,7 @@ The `NonSysId` package introduced in this paper implements an OFR-based system i
 A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO) variant [@guo2015a] of the OFR algorithm. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR, complemented by simulation-based model selection. These enhancements enable robust term selection (compared to the ERR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications, such as fault diagnosis in engineering or the analysis of electrophysiology activity in medical settings, where inputs may not be persistently exciting and separate datasets for validation may be unavailable. 
 `NonSysId` is the only open-sourced package that implements previously proposed solutions to directly address the limitations of the original OFR algorithm. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Moreover, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
 
-## Iterative Orthogonal Forward Regression (iOFR)
+## Iterative Orthogonal Forward Regression (iOFR) with simulation-based model selection
 
 To address the concerns associated with the original OFR, the iOFR algorithm was introduced in [@guo2015a]. To the best of our knowledge, no open-source software implementing this variant currently exists. In the original OFR, the term selection is heavily influenced by the order of orthogonalization, which can often result in incorrect terms being selected in the early stages [@guo2015a;@Mao1997]. Additionally, the order in which terms are selected in the OFR determines the orthogonalization path, resulting in a tree structure of possible models [@guo2015a;@Mao1997]. Finding a globally optimal solution would require an exhaustive search through all orthogonalization paths - an infeasible task given the factorial  growth in paths ($k!$ paths for $k$ terms). The iOFR algorithm addresses this limitation by iteratively exploring multiple orthogonalization paths and re-selecting terms to approximate a globally optimal model without exhaustive search [@guo2015a]. Therefore, it can iteratively produce more globally optimal model structures [@guo2015a]. This approach enables the recovery of correct terms (candidate terms essential for accurately reconstructing dynamics of the original system [@mendes1998a;@AGUIRRE1994]) that might have been overlooked in earlier iterations. As a result, the iOFR generates several candidate models for consideration. 
 
@@ -99,7 +99,7 @@ The model must generalise effectively to unseen data during training, striking a
 
 This section outlines the comprehensive procedures implemented in the `NonSysId` package for identifying (N)ARX models from system input-output data using $\text{iOFR}_{S}$ algorithm combined with PRESS-statistic-based term selection. Additionally, techniques for reducing computational time in $\text{iOFR}_{S}$ are discussed. These techniques focus on efficiently reducing the number of candidate terms, pre-select terms, or both.       
 
[…]
