### 2025-07-10 a99844f
Update paper.md
