+++ paper/paper.md
@@ -101,7 +101,7 @@ This section outlines the comprehensive procedures implemented in the `NonSysId`
 
 Let $\mathcal{D'} = Y \cup U$ be the set of candidate linear terms from past outputs $Y$ and inputs $U$ for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the number of terms in $\mathcal{D'}$ and nonlinearity increases, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity. Likewise, minimizing redundant terms in the initial pre-select set $\mathcal{P}$ can speed up convergence toward an optimum model [@guo2015a]. This section discusses how to effectively reduce $\mathcal{D}''$ and optimize $\mathcal{P}$ for the first iteration to improve $\text{iOFR}_S$ efficiency.
 
-A technique for obtaining a reduced set of candidate model terms, $\mathcal{D}''_{R}$, was proposed in [@Wei2004]. This approach is based on the idea that if a lagged term significantly influences the output of a nonlinear system, it will also be significant in a linearised representation of the system. Accordingly, a linear ARX model is identified first, serving as a linearised model of the actual nonlinear system. The terms from this ARX model are then used to construct $\mathcal{D}''_{R}$. This method has been incorporated into the `NonSysId` package. Regarding the set $\mathcal{P}$, an initialisation method for the iOFR algorithm was proposed in [@guo2015a]. In this method, an overfitting NARX model, $\overline{m_{0}''}$, is first identified using the OFR. The terms from $\overline{m_{0}''}$ are then used to construct the initial set $\mathcal{P}$ for the first iteration of the iOFR. While $\overline{m_{0}''}$ may be sub-optimal, it is likely to include some correct terms. Consequently, using the terms of $\overline{m_{0}''}$ to form the initial set $\mathcal{P}$ ensures fewer redundant terms compared to directly setting $\mathcal{P} \subseteq \mathcal{D}$ [@guo2015a]. 
+A method to obtain a set $\mathcal{D}''{R}$ was proposed in [@Wei2004], based on the idea that important lagged terms in a nonlinear system will also appear in its linearised ARX model. This is implemented in the `NonSysId` package. For initializing the set $\mathcal{P}$ for the first iteration, [@guo2015a] suggests identifying an overfitted NARX model $\overline{m_{0}''}$ using OFR. The terms from $\overline{m_{0}''}$ are then used to form $\mathcal{P}$ for the first iOFR iteration. Although $\overline{m_{0}''}$ may not be optimal, it likely contains some correct terms and results in fewer redundant terms than choosing $\mathcal{P} \subseteq \mathcal{D}''$ directly.
 
 The `NonSysId` package incorporates both aforementioned methods to reduce the computational time of $\text{iOFR}_{S}$. Additionally, the `NonSysId` package implements two new methods, proposed in this paper, to further enhance computational efficiency (referred to as reducing computational time, RCT, methods). \autoref{fig:alg} outlines the procedures of the complete system identification methodology, integrating $\text{iOFR}_{S}$ with these four RCT methods. A brief overview of each RCT method is provided below.
 
