### 2025-07-10 e36a349
Update paper.md
