### 2025-07-10 8200b5e
Update paper.md
