+++ paper/paper.md
@@ -59,7 +59,7 @@ The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regress
 
 1. OFR-ERR may select redundant or incorrect model terms, especially in the presence of complex noise structures or certain input signals.
 
-2. The model structures produced from OFR can be sensitive to the first term selected in the forward selection of model terms.
+2. The model structures produced from OFR can be sensitive to the first term selected in the forward selection of model terms. The term selection is heavily influenced by the order of orthogonalization, which can often result in incorrect terms being selected in the early stages.
 
 3. If the input does not persistently excite the system under consideration (i.e., it lacks the informativeness needed to effectively stimulate the system), the resulting model can be inappropriate. This can result in inaccuracies in long-horizon prediction and, in some cases, even unstable models during simulation (free-run or model-predicted output).
 
@@ -70,12 +70,12 @@ Beyond obtaining parsimonious models, the model should generalise well to unseen
 # Features in `NonSysId`
 
 The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the latter part of the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. 
-A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO) variant [@guo2015a] of the OFR algorithm. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR, complemented by simulation-based model selection. These enhancements enable robust term selection (compared to the ERR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications, such as fault diagnosis in engineering or the analysis of electrophysiology activity in medical settings, where inputs may not be persistently exciting and separate datasets for validation may be unavailable. 
-`NonSysId` is the only open-sourced package that implements previously proposed solutions to directly address the limitations of the original OFR algorithm. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Moreover, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
+
[…]
