+++ paper/paper.md
@@ -89,7 +89,7 @@ In the original iOFR algorithm [@guo2015a], model selection was based on the one
 
 5. Repeat steps 1–4 and iteratively search through different orthogonalization paths.
 
-Steps 2 and 3 are the additions in $\text{iOFR}_{S}$ incorporating simulation-based model selection. In step 2, each model $\Tilde{m}_i \in \Tilde{\mathcal{M}}$, $i=1,\dots,p$, is tested using two inputs: (i) a sequence of 0's, $u^{[0]}(t) = 0 \ \forall t$, and (ii) a sequence of 1's, $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs, $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$, must meet stability conditions for $\Tilde{m}_i$ to be included in $\mathcal{M}$. In this context, stability implies that the outputs remain bounded over time, i.e. stable around a mean without exhibiting exponential growth. Specifically, the responses ($j = 0 \ \text{or} \ 1$) should be around a mean, $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$, with a small variance, $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$. Typically, $\varepsilon = 10^{-2}$. Specifically, for $u^{[0]}(t)$, $\mathbb{E}[\hat{y}^{[0]}(t)] = \beta$, where $\beta$ is the bias term (DC offset) in the model, with $\beta=0$ indicating the absence of a bias term. In step 3, the Bayesian Information Criterion (BIC) [@Schwarz1978;@Stoica2004] is used to select the optimal model $\overline{m}$ from $\mathcal{M}$. The BIC is calculated based on the simulated error variance (i.e. mean squared simulated error [@Piroddi2003]—MSSE) between the actual output and the model's simulated output. The $\text{iOFR}_{S}$ can be represented in functional form as $( \mathcal{M}, \overline{m} ) = \text{iOFR}_{S}( \mathcal{D}, \mathcal{P}, \mathbf{Y})$.
+Steps 2 and 3 in $\text{iOFR}_{S}$ incorparates simulation-based model selection. In Step 2, each candidate model $\Tilde{m}_i \in \Tilde{\mathcal{M}}$, $i=1,\dots,p$ is tested for stability using two constant inputs: $u^{[0]}(t) = 0 \ \forall t$ and $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$ must remain bounded with a small variance, i.e., $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$ and $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$, typically with $\varepsilon = 10^{-2}$. For $u^{[0]}(t)$, the mean equals the model’s bias term $\beta$. In Step 3, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004] selects the best model $\overline{m} \in \mathcal{M}$ based on the mean squared simulated error (MSSE) [@Piroddi2003]. In the latter of this paper, $\text{iOFR}_{S}$ is represented in functional form as $( \mathcal{M}, \overline{m} ) = \text{iOFR}_{S}( \mathcal{D}, \mathcal{P}, \mathbf{Y})$.
 
 ## $\text{iOFR}_{S}$ with reduced computational time
 
@@ -191,7 +191,6 @@ The real data in this example is obtained from an electromechanical system descr
[…]
