+++ paper/paper.md
@@ -73,7 +73,7 @@ The `NonSysId` package introduced in this paper implements an OFR-based system i
 
 A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO) variant [@guo2015a] of the OFR algorithm. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR, complemented by simulation-based model selection. To address the concerns associated with the original OFR, the iOFR algorithm was introduced in [@guo2015a] and to the best of our knowledge, no open-source software implementing this variant currently exists. The iOFR algorithm iteratively explores multiple orthogonalization paths and re-selects terms to approximate a globally optimal model without exhaustive search [@guo2015a]. Therefore, it can iteratively produce more globally optimal model structures [@guo2015a]. This approach enables the recovery of correct terms (candidate terms essential for accurately reconstructing dynamics of the original system [@mendes1998a;@AGUIRRE1994]) that might have been overlooked in earlier iterations. As a result, the iOFR generates several candidate models for consideration. In the original iOFR algorithm [@guo2015a], model selection was based on the one-step-ahead prediction. The implementation of iOFR in the `NonSysId' extends this by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and improve long-horizon prediction accuracy. To ensure good generalisation and obtain a good bias-variance compromise, a PRESS-statistic-based [@Allen1974] term selection with leave-one-out cross-validation was integrated into OFR in [@WANG1996; @hong2003]. Leveraging the orthogonalization procedures in OFR, one-step ahead leave-one-out errors can be computed efficiently, allowing forward selection of terms that minimise the cross-validation error without needing extra validation data. Incorporating this into $\text{iOFR}_{S}$ improves the robustness and generalisation of the selected model terms. 
 
-The above-mentioned enhancements enable robust term selection (compared to the ERR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications, such as fault diagnosis in engineering or the analysis of electrophysiology activity in medical settings, where inputs may not be persistently exciting and separate datasets for validation may be unavailable. `NonSysId` is the only open-sourced package that implements previously proposed solutions to directly address the limitations of the original OFR algorithm. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Moreover, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
[…]
