### 2025-07-10 86b8990
Update paper.md
