+++ paper/paper.md
@@ -75,7 +75,7 @@ A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO
 
 The above-mentioned enhancements enable robust term selection (compared to the ERR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications, such as fault diagnosis in engineering or the analysis of electrophysiology activity in medical settings, where inputs may not be persistently exciting and separate datasets for validation may be unavailable. `NonSysId` is the only open-sourced package that implements previously proposed solutions to directly address the limitations of the original OFR algorithm. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Also, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
 
-## Iterative Orthogonal Forward Regression (iOFR) with simulation-based model selection
+## Iterative Orthogonal Forward Regression (iOFR) with simulation-based model selection ($\text{iOFR}_{S}$)
 
 In the original iOFR algorithm [@guo2015a], model selection was based on the one-step-ahead prediction. The implementation of iOFR in the `NonSysId' extends this by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and improve long-horizon prediction accuracy. Given a vector $\mathbf{Y}$ containing output data, a set of candidate terms $\mathcal{D}$ and a set of pre-select terms $\mathcal{P} \subseteq \mathcal{D}$, where $\mathcal{P} = \{ \phi_1 , \dots, \phi_p  \}$, 1), the procedures for $\text{iOFR}_{S}$ are as follows:
 
@@ -91,11 +91,7 @@ In the original iOFR algorithm [@guo2015a], model selection was based on the one
 
 Steps 2 and 3 in $\text{iOFR}_{S}$ incorparates simulation-based model selection. In Step 2, each candidate model $\Tilde{m}_i \in \Tilde{\mathcal{M}}$, $i=1,\dots,p$ is tested for stability using two constant inputs: $u^{[0]}(t) = 0 \ \forall t$ and $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$ must remain bounded with a small variance, i.e., $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$ and $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$, typically with $\varepsilon = 10^{-2}$. For $u^{[0]}(t)$, the mean equals the model’s bias term $\beta$. In Step 3, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004] selects the best model $\overline{m} \in \mathcal{M}$ based on the mean squared simulated error (MSSE) [@Piroddi2003]. In the latter of this paper, $\text{iOFR}_{S}$ is represented in functional form as $( \mathcal{M}, \overline{m} ) = \text{iOFR}_{S}( \mathcal{D}, \mathcal{P}, \mathbf{Y})$.
 
-## $\text{iOFR}_{S}$ with reduced computational time
-
[…]
