+++ paper/paper.md
@@ -71,25 +71,11 @@ Beyond obtaining parsimonious models, the model should generalise well to unseen
 
 The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the latter part of the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. 
 
-A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO) variant [@guo2015a] of the OFR algorithm. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR, complemented by simulation-based model selection. To address the concerns associated with the original OFR, the iOFR algorithm was introduced in [@guo2015a] and to the best of our knowledge, no open-source software implementing this variant currently exists. The iOFR algorithm iteratively explores multiple orthogonalization paths and re-selects terms to approximate a globally optimal model without exhaustive search [@guo2015a]. Therefore, it can iteratively produce more globally optimal model structures [@guo2015a]. This approach enables the recovery of correct terms (candidate terms essential for accurately reconstructing dynamics of the original system [@mendes1998a;@AGUIRRE1994]) that might have been overlooked in earlier iterations. As a result, the iOFR generates several candidate models for consideration. In the original iOFR algorithm [@guo2015a], model selection was based on the one-step-ahead prediction. The implementation of iOFR in the `NonSysId' extends this by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and improve long-horizon prediction accuracy. To ensure good generalisation and obtain a good bias-variance compromise, a PRESS-statistic-based [@Allen1974] term selection with leave-one-out cross-validation was integrated into OFR in [@WANG1996; @hong2003]. Leveraging the orthogonalization procedures in OFR, one-step ahead leave-one-out errors can be computed efficiently, allowing forward selection of terms that minimise the cross-validation error without needing extra validation data. Incorporating this into $\text{iOFR}_{S}$ improves the robustness and generalisation of the selected model terms. 
[…]
