+++ paper/paper.md
@@ -77,11 +77,11 @@ The above-mentioned enhancements enable robust term selection (compared to the E
 
 In the original iOFR algorithm [@guo2015a], model selection was based on the one-step-ahead prediction. The implementation of iOFR in the `NonSysId' extends this by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and improve long-horizon prediction accuracy. There are two simple tests done to ensure simulation stability (bounded-input bounded-output). Using a pre-select term list $\mathcal{P}$, the iFRO algorithm explores several orthogonalization paths each producing a candidate model. These models are tested for stability using two constant inputs: $u^{[0]}(t) = 0 \ \forall t$ and $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$ must remain bounded with a small variance, i.e., $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$ and $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$ ($j = 0 \text{or} 1$), typically with $\varepsilon = 10^{-2}$. For $u^{[0]}(t)$, the mean equals the model’s bias (DC-offset) term $\beta$. Afterwards, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004], based on the mean squared simulated error (MSSE) [@Piroddi2003], selects the best model $\overline{m}$ from the previously determined set of stable models. $\mathcal{P}$ is updated with the term of $\overline{m}$, and the next iteration begins.
 
-Let $\mathcal{D'} = Y \cup U$ be the set of candidate linear terms from past outputs $Y$ and inputs $U$ for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the number of terms in $\mathcal{D'}$ and nonlinearity increases, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity. Likewise, minimizing redundant terms in the initial pre-select set $\mathcal{P}$ can speed up convergence toward an optimum model [@guo2015a]. This section discusses how to effectively reduce $\mathcal{D}''$ and optimize $\mathcal{P}$ for the first iteration to improve $\text{iOFR}_S$ efficiency.
+Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity considered increases, $|\mathcal{D}''|$ can grow exponentially. On top of this, if the number of past outputs and inputs considered also increases, $|\mathcal{D}''|$ explodes [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity. Likewise, minimizing redundant terms in the initial pre-select set $\mathcal{P}$ ($\mathcal{P}$ at the first iteration) can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements four strategies to effectively reduce $\mathcal{D}''$ ($\mathcal{D}''_R$) and optimize the initial set $\mathcal{P}$ to improve $\text{iOFR}_S$ efficiency.
 
[…]
