### 2025-07-10 69f7c06
Update paper.md
