+++ paper/paper.md
@@ -79,9 +79,7 @@ In the original iOFR algorithm [@guo2015a], model selection was based on the one
 
 Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity considered increases, $|\mathcal{D}''|$ can grow exponentially. On top of this, if the number of past outputs and inputs considered also increases, $|\mathcal{D}''|$ explodes [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity. Likewise, minimizing redundant terms in the initial pre-select set $\mathcal{P}$ ($\mathcal{P}$ at the first iteration) can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements four strategies to effectively reduce $\mathcal{D}''$ ($\mathcal{D}''_R$) and optimize the initial set $\mathcal{P}$ to improve $\text{iOFR}_S$ efficiency.
 
-A strategy to obtain a set $\mathcal{D}''_{R}$ was proposed in [@Wei2004], based on the idea that important lagged terms in a nonlinear system will also appear in its linearised ARX model. This is implemented in the `NonSysId` package. For the initial $\mathcal{P}$, [@guo2015a] suggests identifying an overfitted NARX model $\overline{m_{0}''}$ using OFR. The terms from $\overline{m_{0}''}$ are then used to form $\mathcal{P}$ for the first iOFR/$\text{iOFR}_S$ iteration. Although $\overline{m_{0}''}$ may not be optimal, it likely contains some correct terms and results in fewer redundant terms than choosing $\mathcal{P} \subseteq \mathcal{D}''$ directly.
-
-The `NonSysId` package incorporates both aforementioned strategies to reduce the computational time of $\text{iOFR}_{S}$. Additionally, the `NonSysId` package implements two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as reducing computational time, RCT, methods). \autoref{fig:alg} outlines the procedures of the complete system identification methodology, integrating $\text{iOFR}_{S}$ with these four RCT methods. A brief overview of each RCT method is provided below.
+A strategy to obtain a set $\mathcal{D}''_{R}$ was proposed in [@Wei2004], based on the idea that important lagged terms in a nonlinear system will also appear in its linearised ARX model. This is implemented in the `NonSysId` package. For the initial $\mathcal{P}$, [@guo2015a] suggests identifying an overfitted NARX model $\overline{m_{0}''}$ using OFR. The terms from $\overline{m_{0}''}$ are then used to form $\mathcal{P}$ for the first iOFR/$\text{iOFR}_S$ iteration. Although $\overline{m_{0}''}$ may not be optimal, it likely contains some correct terms and results in fewer redundant terms than choosing $\mathcal{P} \subseteq \mathcal{D}''$ directly. The `NonSysId` package incorporates both aforementioned strategies to reduce the computational time of $\text{iOFR}_{S}$. Additionally, the `NonSysId` package implements two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as strategies reducing computational time, RCT). A brief overview of each RCT strategy is provided below.
 
 - **RCT Method 1**: This method, as proposed in [@Wei2004], seeks to obtain a reduced set of candidate model terms, enabling $\text{iOFR}_{S}$ to operate within a narrower search space defined by $\mathcal{D}''_{R} \subset \mathcal{D}''$.
 
[…]
