### 2025-07-11 05dffc4
Update paper.md
