+++ paper/paper.md
@@ -30,15 +30,13 @@ System identification involves constructing mathematical models of dynamic syste
 
 # Statement of Need
 
-System identification is a field at the intersection of control theory, dynamic systems theory and machine learning that seeks to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Generally, system identification has two primary objectives [@ljung1998system;@billings2013a]: (i) to accurately map the system's inputs and outputs, allowing for the prediction of new, unseen data, and (ii) to capture the underlying dynamics of the system within the model. The dynamic models generated through system identification can be either discrete-time or continuous-time models [@UNBEHAUEN1997]. This paper centres on widely-used discrete-time models, specifically nonlinear auto-regressive models with exogenous inputs (NARX), where the ARX model is a linear variant of the NARX structure. 
-
-These input-output time-series models predict future outputs of a system based on its historical input and output instances. Therefore,  system identification is employed to determine a specific functional relationship that maps past input instances (input-lagged terms) and past output instances (output-lagged terms), to the present output instance in time $y(t)$. $t$ here refers to a time index (i.e. $t$^th^ sample). This functional mapping can take the form of a polynomial, a neural network, or even a fuzzy logic-based model. Here, we focus on polynomial NARX models with a maximum polynomial degree $N_p \in \mathbb{Z}^{+}$. An example of a polynomial NARX model can be
+System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models (can be continuous-time or discrete-time models [@UNBEHAUEN1997]) of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) to capture the underlying dynamics of the system within the model [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. These are input-output time-series models that predict a system’s future outputs based on its past inputs and outputs. Therefore, system identification determines a functional mapping from past input/output terms (lagged terms) to the current output $y(t)$, where $t$ denotes the time index. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models with maximum degree $N_p \in \mathbb{Z}^{+}$. For example:
 
 \begin{equation}\label{eq:narx_exmpl}
     y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
 \end{equation}
 
-In this example, $\phi_{1}(t)=y(t-1)$ and $\phi_{2}(t)=u(t-2)$ have a degree of 1 and are the linear terms (1\textsuperscript{st} order monomials or linear monomials) of the model. $\phi_{3}(t) = y(t-2)^{2}u(t-1)^{3}$ is a nonlinear term with a degree of $5$ (5^th^ order monomial, more generally a nonlinear monomial). The NARX model given in \autoref{eq:narx_exmpl} has a polynomial degree $N_p=5$ (highest degree of any monomial). NARX models have been applied extensively to model and analyse complex systems in fields such as control, condition monitoring, advanced manufacturing and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Moreover, it has been demonstrated that the NARX model is equivalent to a recurrent neural network (RNN) [@Sum1999]. Extending the NARX model to incorporate a noise model, we obtain the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model. 
[…]
