### 2025-07-11 eba568c
Update paper.md
