+++ paper/paper.md
@@ -38,9 +38,7 @@ System identification is at the intersection of control theory, dynamic systems 
 
 Here, $\phi_{1}(t)=y(t-1)$ and $\phi_{2}(t)=u(t-2)$ are linear terms (degree $1$ monomials or linear monomials), while$\phi_{3}(t) = y(t-2)^{2}u(t-1)^{3}$ is nonlinear term (degree $5$ monomial, more generally a nonlinear monomial). The NARX model given in \autoref{eq:narx_exmpl} has a polynomial degree $N_p=5$ (highest degree of any monomial). NARX models are widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Notably, they are equivalent to RNNs [@Sum1999]. Extending the NARX model with a noise model yields the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model.
 
-The primary challenge in learning a polynomial NARX model is to identify the polynomial structure of the model, i.e. selecting which terms from a set of candidate model terms (monomials) should be included in the model. However, determining the appropriate linear and nonlinear terms to include in the model structure is critical to achieving parsimonious models. This is particularly important in the nonlinear cases (Chapter 1, [@billings2013a]), as the inclusion of  unnecessary model terms can result in a model that erroneously captures dynamics that do not belong to the underlying system [@AGUIRRE1995;@mendes1998a]. The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regression OLS (FROLS) [@chen1989b;@billings1987a], is based on the Orthogonal Least Squares (OLS). When combined with an appropriate term selection criterion [@korenberg1988a;@WANG1996;@hong2003], it efficiently selects model terms (regressors) in a forward selection manner (sequentially according to the term selection criterion), facilitating the development of a parsimonious model. The OFR/FROLS algorithm evaluates (according to a term selection criterion) the impact of each term on the output independently of the influence of other terms, which is achieved through  orthogonalization procedures. This allows for the sequential inclusion of appropriate terms in the final model using a forward selection approach. The most commonly used and widely accepted model term selection criterion for the OFR algorithm is the error reduction ratio (ERR). During the forward selection procedures, the ERR selects the term that maximises explained variance, thereby  maximising the goodness of fit. Over the years, many variants of the OFR have been proposed. 
-
-Recently, two open-sourced packages for system identification have been introduced: `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive. However, they are based on the original FROLS/OFR algorithm. Concerns persist regarding the original OFR algorithm (OFR-ERR), which relies on the ERR for term selection [@Piroddi2003;@Mao1997], for example,
[…]
