+++ paper/paper.md
@@ -48,7 +48,7 @@ The key challenge in learning a polynomial NARX model is to identify an appropri
 
 4. The ERR focuses solely on explained variance when selecting terms, which can lead to overfitting.
 
-Beyond obtaining parsimonious models, it is essential that models generalise well to unseen data--validation data not used during training (model identification). This reflects the bias–variance trade-off, often addressed through suitable cross-validation strategies [@Little2017; @Stone1974]. However, in some applications, such as structural health monitoring or fault diagnosis, obtaining separate validation data may be impractical [@Gharehbaghi2022; @Vamsikrishna2024]. Similarly, in neuroscience, rapidly changing dynamics between brain regions can occur within milliseconds, making it difficult to obtain electrophysiological data that precisely capture the exact same behaviour [@Kunjan2021; @Seedat2024; @Chen2016; @Eichenbaum2021; @Lehnertz2021]. These challenges are critical when applying system identification to specific domains. The following section outlines the features in the `NonSysId` package, designed to address these issues.
+Beyond obtaining parsimony, models must generalise well to unseen data--validation data not used during training (model identification). This reflects the bias–variance trade-off, often addressed through suitable cross-validation strategies [@Little2017; @Stone1974]. However, in some applications, such as structural health monitoring or fault diagnosis, obtaining separate validation data may be impractical [@Gharehbaghi2022; @Vamsikrishna2024]. Similarly, in neuroscience, rapidly changing complex dynamics between brain regions can occur within milliseconds, making it difficult to obtain electrophysiological data that precisely capture the same behaviour for validation purposes [@Kunjan2021; @Seedat2024; @Chen2016; @Eichenbaum2021; @Lehnertz2021]. These challenges are critical when applying system identification to specific domains. The following section outlines the features in the `NonSysId` package, designed to address these issues.
 
 # Features in `NonSysId`
 
