### 2025-07-11 02b7384
Update paper.md
