+++ paper/paper.md
@@ -38,7 +38,7 @@ System identification is at the intersection of control theory, dynamic systems 
 
 Here, $\phi_{1}(t)=y(t-1)$ and $\phi_{2}(t)=u(t-2)$ are linear terms (degree $1$ monomials or linear monomials), while$\phi_{3}(t) = y(t-2)^{2}u(t-1)^{3}$ is nonlinear term (degree $5$ monomial, more generally a nonlinear monomial). The NARX model given in \autoref{eq:narx_exmpl} has a polynomial degree $N_p=5$ (highest degree of any monomial). NARX models are widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Notably, they are equivalent to RNNs [@Sum1999]. Extending the NARX model with a noise model yields the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model.
 
-The key challenge in learning a polynomial NARX model is to identify an appropriate polynomial structure, i.e. from a set of candidate terms (monomials), selecting which terms to include. Choosing the appropriate linear and nonlinear terms is vital for producing parsimonious models, especially in nonlinear cases (Chapter 1, [@billings2013a]). Including superfluous terms can lead to spurious dynamics unrelated to the true system behaviour [@AGUIRRE1995;@mendes1998a]. The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regression OLS (FROLS) [@chen1989b; @billings1987a], is based on Orthogonal Least Squares (OLS). When paired with a suitable term selection criterion [@korenberg1988a; @WANG1996; @hong2003], it efficiently selects regressors via forward selection. Through orthogonalization, OFR evaluates the individual contribution of each term to the output, enabling sequential inclusion of significant terms. The most widely used term selection criterion for the OFR is the error reduction ratio (ERR), which chooses terms that maximise the explained variance during forward selection, thus incrementally improving model fit. Numerous variants of OFR have since been proposed. Recently, two open-sourced packages for system identification have been introduced: `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive. However, they are based on the original FROLS/OFR algorithm. Concerns persist regarding the original OFR algorithm (OFR-ERR), which relies on the ERR for term selection [@Piroddi2003;@Mao1997], for example,
[…]
