### 2025-07-11 f3582bd
Update paper.md
