### 2025-07-11 52a1018
Update paper.md
