+++ paper/paper.md
@@ -60,21 +60,19 @@ To address the concerns associated with the original OFR, the iOFR algorithm was
 
 To ensure good generalisation and obtain a good bias-variance compromise, a PRESS-statistic-based [@Allen1974] term selection with leave-one-out cross-validation was integrated into OFR in [@WANG1996; @hong2003]. Leveraging the orthogonalization procedures in OFR, one-step ahead leave-one-out errors can be computed efficiently, allowing forward selection of terms that minimise the predicted cross-validation error without extra validation data. Incorporating this into $\text{iOFR}_{S}$ improves the robustness and generalisation of the selected model terms. 
 
-Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity. Likewise, minimizing redundant terms in $\mathcal{P}$ at the first iteration can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements four strategies to effectively reduce $\mathcal{D}''$ ($\mathcal{D}''_R$) and optimize the initial set $\mathcal{P}$ to improve $\text{iOFR}_S$ efficiency.
+Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing this set to $\mathcal{D}''_R \subset \mathcal{D}''$ can reduce complexity [@Wei2004]. Likewise, minimizing redundant terms in $\mathcal{P}$ at the first iteration can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements four strategies to effectively reduce $\mathcal{D}''$ ($\mathcal{D}''_R$) and optimize the initial set $\mathcal{P}$ to improve $\text{iOFR}_S$ efficiency. Two of which are already suggested in [@Wei2004] and [@guo2015a]. Additionally, the `NonSysId` package implements two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as strategies reducing computational time, RCT). A brief overview of each RCT strategy is provided below.
 
[…]
