### 2025-07-14 1d41875
Update paper.md
