+++ paper/paper.md
@@ -56,15 +56,16 @@ The `NonSysId` package introduced in this paper implements an OFR-based system i
 
 A notable feature of `NonSysId` is the implementation of the iterative OFR (iFRO) variant [@guo2015a] of the OFR algorithm. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR, complemented by simulation-based stability tests and model selection. These enhancements enable robust term selection (compared to the original OFR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications where inputs may not be persistently exciting and separate datasets for validation may be unavailable. `NonSysId` is the only open-sourced package that implements previously proposed solutions to address the limitations of the original OFR algorithm directly. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Also, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
 
-To address the concerns associated with the original OFR, the iOFR algorithm was introduced in [@guo2015a] and to the best of our knowledge, no open-source software implementing this variant currently exists. Using a pre-select term list $\mathcal{P}$, the iFRO algorithm explores several orthogonalization paths (forward selection routes), each producing a candidate model. According to a criterion, based on the one-step ahead prediction, the best model $\overline{m}$ is chosen. $\mathcal{P}$ is updated with the term of $\overline{m}$, and the next iteration begins. As demonstrated in [@guo2015a], each iteration produces more globally optimal model structures than the previous iterations. The implementation of iOFR in the `NonSysId' extends this by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and improve long-horizon (multi-step ahead) prediction accuracy. There are two simple tests done to ensure simulation stability (bounded-input bounded-output). The several candidate models at each iteration of iOFR are tested for stability using two constant inputs: $u^{[0]}(t) = 0 \ \forall t$ and $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$ must remain bounded with a small variance, i.e., $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$ and $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$ ($j = 0 \text{or} 1$) (typically $\varepsilon = 10^{-2}$). $\mathbb{E}[\hat{y}^{[0]}(t)] = \beta$, which is the bias (DC-offset) term of the model. Afterwards, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004], based on the mean squared simulated error (MSSE) [@Piroddi2003], selects the best model $\overline{m}$ from the previously determined set of stable models. 
[…]
