### 2025-07-14 6c8794e
Update paper.md
