+++ paper/paper.md
@@ -62,7 +62,6 @@ To ensure good generalisation and obtain a good bias-variance compromise, a PRES
 
 Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing $\mathcal{D}''$, $\mathcal{D}''_R$ can reduce complexity [@Wei2004]. Likewise, minimizing redundant terms in $\mathcal{P}$ at the first iteration can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements these and two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as strategies reducing computational time, RCT). A brief overview of each RCT strategy is provided below. 
 
-
 - **RCT Method 1**: This method, as proposed in [@Wei2004], seeks to obtain a data-driven $\mathcal{D}''_R$, enabling $\text{iOFR}_{S}$ to operate within a narrower search space.
 
 - **RCT Method 2**: This method, as proposed in [@guo2015a], identifies an appropriate $\mathcal{P}$ for the first iteration of the iOFR algorithm by identifying an overfitted NARX model $\overline{m_{0}''}$ using OFR. Terms of $\overline{m_{0}''}$ are used to form $\mathcal{P}$ as it is likely to contain some correct terms and results in fewer redundant terms than choosing $\mathcal{P} \subseteq \mathcal{D}''$ directly.
@@ -73,7 +72,7 @@ Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and in
 
 Among the methods, RCT 3 is most effective, followed by 1, 4, and 2. However, methods 1 and 3 may miss important terms, especially in noisy or complex systems, leading to sub-optimal models. RCT methods also add overhead; therefore, if $\mathcal{D}''$ is already small, using $\text{iOFR}_{S}$ without RCT ($\mathcal{P} \subseteq \mathcal{D}''$) may be more efficient. The next section includes examples from the `NonSysId` package.
 
-NARX models can be analysed in the frequency domain using Nonlinear Output Frequency Response Functions (NOFRFs) [@Lang2005], which extend classical frequency response analysis to nonlinear systems (Chapter 6, [@billings2013a]). This facilitates a detailed understanding of how nonlinearities affect input-output dynamics [@Lang2005;@BAYMA2018]. Consequently, NOFRFs enhance the utility of NARX models by offering a comprehensive framework for analysing and interpreting nonlinear systems [@Zhu2021].
+NARX models can be analysed in the frequency domain using Nonlinear Output Frequency Response Functions (NOFRFs) [@Lang2005], which extend classical frequency response analysis to nonlinear systems (Chapter 6, [@billings2013a]). This facilitates a detailed understanding of how nonlinearities affect input-output dynamics [@Lang2005;@BAYMA2018;@Gunawardena2018;@ZHU2022]. Consequently, NOFRFs enhance the utility of NARX models by offering a comprehensive framework for analysing and interpreting nonlinear systems [@Zhu2021].
 
 # Examples
 
@@ -87,13 +86,13 @@ The following example demonstrates how to identify a NARX model using the `NonSy
     - 0.1597y(t-1)u(t-1) - 0.1396y(t-1)u(t-2)\\
     + 0.1297y(t-2)u(t-1) + 0.1086y(t-2)u(t-2) + 0.0085y(t-2)^2
 \end{multline}
[…]
