### 2025-07-14 ae407b3
Update paper.md
