### 2025-07-14 b54c2e1
Update paper.md
