### 2025-07-14 1a31f96
Update paper.md
