+++ paper/paper.md
@@ -58,11 +58,11 @@ To address the concerns associated with the original OFR, the iterative OFR (iFR
 
 Using a pre-select term list $\mathcal{P}$, the iFRO algorithm explores several orthogonalization paths (forward selection routes), each producing a candidate model. According to a criterion, based on the one-step ahead prediction, the best model $\overline{m}$ is chosen. $\mathcal{P}$ is updated with the term of $\overline{m}$, and the next iteration begins. As demonstrated in [@guo2015a], each iteration produces more globally optimal model structures than the previous iterations. In `NonSysId', $\text{iOFR}_{S}$ extends the iOFR where two simple tests are done to ensure simulation stability (bounded-input bounded-output). The candidate models evaluated at each iteration of iOFR are tested for stability using two constant inputs: $u^{[0]}(t) = 0 \ \forall t$ and $u^{[1]}(t) = 1 \ \forall t$. The corresponding simulated outputs $\hat{y}^{[0]}(t)$ and $\hat{y}^{[1]}(t)$ must remain bounded with a small variance, i.e., $\mathbb{E}[\hat{y}^{[j]}(t)] \in \mathbb{R}$ and $\text{Var}(\hat{y}^{[j]}(t)) \leq \varepsilon$ ($j = 0 \text{or} 1$) (typically $\varepsilon = 10^{-2}$). $\mathbb{E}[\hat{y}^{[0]}(t)] = \beta$, which is the bias (DC-offset) term of the model. Afterwards, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004], based on the mean squared simulated error (MSSE) [@Piroddi2003], selects the best model $\overline{m}$ from the previously determined set of stable models. 
 
-Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs for ARX model identification. For NARX models, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, so $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing $\mathcal{D}''$, $\mathcal{D}''_R$ can reduce complexity [@Wei2004]. Likewise, minimizing redundant terms in $\mathcal{P}$ at the first iteration can speed up convergence toward an optimum model [@guo2015a]. The `NonSysId` package implements these and two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as strategies reducing computational time, RCT). A brief overview of each RCT strategy is provided below. 
+Let $\mathcal{D'}$ be the set of candidate linear terms from past outputs and inputs. Extending $\mathcal{D'}$, the candidate set $\mathcal{D}''$ includes both linear and nonlinear terms, $\mathcal{D}'' \supset \mathcal{D'}$. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}''|$ can grow exponentially [@billings2013a], increasing computational cost for iOFR/$\text{iOFR}_S$. Appropriately reducing $\mathcal{D}''$, $\mathcal{D}''_R$ [@Wei2004] and minimizing redundant terms in $\mathcal{P}$ at the first iteration [@guo2015a] can speed up convergence toward an optimum model. The `NonSysId` package implements these and two new strategies, proposed in this paper, to further enhance computational efficiency (referred to as strategies reducing computational time, RCT). A brief overview of each RCT strategy is provided below. 
 
[…]
