### 2025-07-14 0a5adcb
Update paper.md
