### 2025-07-14 2fe6a1c
Update paper.md
