### 2025-07-14 72b3ab3
Update paper.md
