### 2025-07-14 5f37ab3
Update paper.md
