### 2025-07-14 a45b1e9
Update paper.md
