+++ paper/paper.md
@@ -26,29 +26,29 @@ bibliography: paper.bib
 ---
 
 # Summary
-System identification involves constructing mathematical models of dynamic systems using input-output data, enabling analysis and prediction of system behaviour in both time and frequency domains. This approach can model the entire system or capture specific dynamics within it. For meaningful analysis, it is essential for the model to accurately reflect the underlying system's behaviour. This paper introduces NonSysId, an open-sourced MATLAB software package designed for nonlinear system identification, specifically focusing on NARMAX models. The software incorporates an advanced term selection methodology that prioritises simulation (free-run) accuracy while preserving model parsimony. A key feature is the integration of iterative Orthogonal Forward Regression (iOFR) with Predicted Residual Sum of Squares (PRESS) statistic-based term selection, facilitating robust model generalisation without the need for a separate validation dataset. Furthermore, techniques for reducing computational overheads are implemented. These features make `NonSysId` particularly suitable for real-time applications such as structural health monitoring, fault diagnosis, and biomedical signal processing, where it is a challenge to capture the signals under consistent conditions, resulting in limited or no validation data.
+System identification involves constructing mathematical models of complex dynamic systems using input-output data, enabling analysis and prediction of system behaviour in both time and frequency domains. This approach can model the entire system or capture specific dynamics within it. For meaningful analysis, it is essential for the model to accurately reflect the underlying system's behaviour. This paper introduces NonSysId, an open-sourced MATLAB software package designed for nonlinear system identification, specifically focusing on NARMAX models. The software incorporates an advanced term selection methodology that prioritises simulation (free-run) accuracy while preserving model parsimony. A key feature is the integration of iterative Orthogonal Forward Regression (iOFR) with Predicted Residual Sum of Squares (PRESS) statistic-based term selection, facilitating robust model generalisation without the need for a separate validation dataset. Furthermore, techniques for reducing computational overheads are implemented. These features make `NonSysId` particularly suitable for real-time applications such as structural health monitoring, fault diagnosis, and biomedical signal processing, where it is a challenge to capture the signals under consistent conditions, resulting in limited or no validation data.
 
 # Statement of Need
 
-System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models (can be continuous-time or discrete-time models [@UNBEHAUEN1997]) of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) to capture the underlying dynamics of the system within the model [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. These are input-output time-series models that predict a system’s future outputs based on its past inputs and outputs. Therefore, system identification determines a functional mapping from past input/output terms (lagged terms) to the current output $y(t)$, where $t$ denotes the time index. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models with maximum degree $N_p \in \mathbb{Z}^{+}$. For example:
[…]
