### 2025-07-14 c389752
Update paper.md
