### 2025-07-15 b131a46
Update paper.md
