### 2025-07-15 3f9bc8b
Update paper.md
