+++ paper/paper.md
@@ -54,11 +54,11 @@ Beyond obtaining parsimony, models must generalise well to unseen data--validati
 
 The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. 
 
-To address the concerns associated with the original OFR, the iterative OFR (iFRO) algorithm was introduced in [@guo2015a], which, to the best of our knowledge, is not implemented in any open-source software. The implementation of iOFR in the `NonSysId` extends it by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and further improve long-horizon (multi-step ahead) prediction accuracy. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR/$\text{iOFR}_{S}$. Leveraging the orthogonalization procedures in OFR, one-step ahead leave-one-out errors [@Allen1974] can be computed efficiently, allowing forward selection of terms that minimise the predicted cross-validation error without extra validation data [@WANG1996; @hong2003]. Incorporating this into $\text{iOFR}_{S}$ improves the robustness and generalisation of the evaluated model terms--obtaining a good bias-variance compromise. These enhancements enable robust term selection (compared to the original OFR), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications where inputs may not be persistently exciting and separate validation datasets may be unavailable. `NonSysId` is the only open-sourced package that implements previously proposed solutions to address the limitations of the original OFR algorithm directly. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate term set, significantly speeding up the forward selection process. Also, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983].
[…]
