### 2025-07-15 3d1aefe
Update paper.md
