### 2025-07-16 e98e0f5
Update paper.md
