### 2025-07-16 f1c380f
Update paper.md
