### 2025-07-16 8f5f2c4
Update paper.md
