### 2025-07-16 7c3f0fc
Update paper.md
