+++ paper/paper.md
@@ -44,9 +44,9 @@ Beyond obtaining parsimony, models must generalise well to unseen data, i.e. obt
 
 # Features in `NonSysId`
 
-The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. 
+The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. At the time of writing, `NonSysId` is the only open-sourced package that implements previously proposed solutions to address the limitations of the original OFR algorithm directly. 
 
-To address the concerns associated with the original OFR, the iterative OFR (iFRO) algorithm was introduced in [@guo2015a], which, to the best of our knowledge, is not implemented in any open-source software. The implementation of iOFR in the `NonSysId` extends it by incorporating simulation-based model selection ($\text{iOFR}_{S}$) to ensure simulation stability and further improve long-horizon (multi-step ahead) prediction accuracy. Additionally, the PRESS-statistic-based term selection [@WANG1996;@hong2003] is integrated with the iOFR/$\text{iOFR}_{S}$. Leveraging orthogonalization, one-step ahead leave-one-out errors [@Allen1974] can be computed efficiently, allowing step-wise forward selection of terms that minimise the predicted cross-validation error without extra validation data [@WANG1996; @hong2003]. These enhancements enable robust term selection (compared to the original OFR-EER), built-in cross-validation, and the ability to produce models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. With these features, the `NonSysId` package makes system identification feasible for real-time applications where inputs may not be persistently exciting and separate validation datasets may be unavailable. `NonSysId` is the only open-sourced package that implements previously proposed solutions to address the limitations of the original OFR algorithm directly. For NARX models, where the candidate term set can be extensive and computationally demanding in the iFRO algorithm [@chen1989b; @billings1987a; @guo2015a], `NonSysId` incorporates methods to reduce the candidate term set [@Wei2004], significantly speeding up the forward selection process. Also, the package includes correlation-based residual analysis techniques for nonlinear model validation [@Billings1983]. 
+To overcome limitations of the original OFR, the NonSysId package implements the iterative OFR (iOFR) algorithm [@guo2015a], extended with simulation-based model selection ($\text{iOFR}_{S}$) for improved long-horizon prediction and stability. It integrates the PRESS-statistic-based [@Allen1974] term selection for OFR [@WANG1996; @hong2003], enabling efficient step-wise forward selection of terms that minimise the one-step ahead leave-one-out cross-validation error. These enhancements allow for robust term selection (compared to the original OFR-EER), built-in cross-validation, and models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. Therefore, the package makes system identification feasible for real-time applications where inputs may not be persistently exciting and separate validation datasets may be unavailable. For NARX models, the candidate term set can be extensive and computationally demanding in the iFRO algorithm [@chen1989b; @billings1987a; @guo2015a]. `NonSysId`, based on [@Wei2004] and [@guo2015a], incorporates strategies to speed up the iterative step-wise forward selection process. Additionally, the package includes correlation-based residual analysis techniques for validating nonlinear models [@Billings1983].  
 
[…]
