+++ paper/paper.md
@@ -38,9 +38,9 @@ System identification is at the intersection of control theory, dynamic systems 
 
 Here, $y(t-1)$ and $u(t-2)$ are linear terms (degree $1$ monomials), while $y(t-2)^{2}u(t-1)^{3}$ is a nonlinear term (degree $5$ monomial). The NARX model given in \autoref{eq:narx_exmpl} has a polynomial degree $N_p=5$ (highest degree of any monomial). NARX models are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model. Additionally, Nonlinear Output Frequency Response Functions (NOFRFs) [@Lang2005] facilitate a detailed frequency-domain understanding of how nonlinearities affect input-output dynamics captured by NARX models [@Lang2005;@BAYMA2018;@Gunawardena2018;@ZHU2022]. This enhances the utility of NARX models by offering a comprehensive framework for analysing and interpreting nonlinear systems [@Zhu2021].
 
-The key challenge in learning a polynomial NARX model is to identify a parsimonious model structure from a set of candidate terms (monomials). Especially in nonlinear cases (Chapter 1, [@billings2013a]), superfluous terms can lead to spurious dynamics [@AGUIRRE1995;@mendes1998a]. The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regression OLS (FROLS) [@chen1989b; @billings1987a], based on Orthogonal Least Squares (OLS), efficiently selects regressors. Through orthogonalization, according to a term selection criterion [@korenberg1988a; @WANG1996; @hong2003], the OFR can evaluate the individual contribution of each term to the output, enabling step-wise forward selection of terms. Using the error reduction ratio (ERR), a widely used term selection criterion, the OFR incrementally chooses terms that maximise the explained variance (model fit).  Numerous variants of OFR have since been proposed. Recently, two open-sourced packages for system identification have been introduced: `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive. However, they are based on the original FROLS/OFR algorithm. Concerns persist regarding the original OFR algorithm (OFR-ERR), which relies on the ERR for term selection [@Piroddi2003;@Mao1997]. These concerns are mainly regarding the selection of redundant/incorrect terms, overfitting and inaccuracies/instability in long-horizon prediction (simulation, free-run or model-predicted output). These are due to inputs which do not persistently excite the system under consideration (i.e., it lacks the informativeness needed to effectively stimulate the system) or the presence of complex noise structures. 
[…]
