### 2025-07-16 d593833
Update paper.md
