### 2025-07-16 3cd5ead
Update paper.md
