+++ paper/paper.md
@@ -30,15 +30,15 @@ System identification involves constructing mathematical models of complex dynam
 
 # Statement of Need
 
-System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. In this case, system identification determines a functional mapping between past input/output terms (input/output lagged terms) to the current output $y(t)$, where $t$ denotes the time index. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models with maximum degree $N_p \in \mathbb{Z}^{+}$. For example:
+System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. In this setting, system identification determines a functional mapping between past input/output terms (input/output lagged terms) to the current output $y(t)$, where $t$ denotes the time index. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models with maximum degree $N_p \in \mathbb{Z}^{+}$. For example:
 
 \begin{equation}\label{eq:narx_exmpl}
     y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
 \end{equation}
 
-Here, $y(t-1)$ and $u(t-2)$ are linear terms (degree $1$ monomials), while $y(t-2)^{2}u(t-1)^{3}$ is a nonlinear term (degree $5$ monomial). The NARX model given in \autoref{eq:narx_exmpl} has a polynomial degree $N_p=5$ (highest degree of any monomial). NARX models are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model. Additionally, Nonlinear Output Frequency Response Functions (NOFRFs) [@Lang2005] facilitate a detailed frequency-domain understanding of how nonlinearities affect input-output dynamics captured by NARX models [@Lang2005;@BAYMA2018;@Gunawardena2018;@ZHU2022]. This enhances the utility of NARX models by offering a comprehensive framework for analysing and interpreting nonlinear systems [@Zhu2021].
[…]
