### 2025-07-16 7ff600a
Update paper.md
