### 2025-07-17 c1d3c70
Update paper.md
