### 2025-07-17 c8d25e2
Update paper.md
