+++ paper/paper.md
@@ -46,11 +46,7 @@ Beyond obtaining parsimony, models must generalise well to unseen data, which is
 
 The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues mentioned in the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], along with a proposed simulation-based model selection procedure. At the time of writing, `NonSysId` is the only open-sourced package that implements previously proposed solutions to address the limitations of the original OFR algorithm directly. 
 
-To overcome limitations of the original OFR, the `NonSysId` package implements the iterative OFR (iOFR) algorithm [@guo2015a], extended with simulation-based model selection ($\text{iOFR}_{S}$) for improved long-horizon prediction and stability. It integrates the PRESS-statistic-based [@Allen1974] term selection for OFR [@WANG1996; @hong2003], enabling efficient step-wise forward selection of terms that minimise the one-step ahead leave-one-out cross-validation error. These enhancements allow for robust term selection (compared to the original OFR-EER), built-in cross-validation, and generate models with long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. Therefore, the package makes system identification feasible for real-time applications where inputs may not be persistently exciting and separate validation datasets may be unavailable. For NARX models, the candidate term set can be extensive and computationally demanding in the iFRO algorithm [@chen1989b; @billings1987a; @guo2015a]. `NonSysId`, based on [@Wei2004] and [@guo2015a], incorporates strategies to speed up the iterative step-wise forward selection process. Additionally, the package includes correlation-based residual analysis techniques for validating nonlinear models [@Billings1983].  
-
-Using a pre-select term list $\mathcal{P}$, the iFRO algorithm explores several orthogonalization paths (forward selection routes), each producing a candidate model. According to a criterion based on one-step-ahead prediction, the best model $\overline{m}$ is chosen. $\mathcal{P}$ is updated with the terms of $\overline{m}$, and the next iteration begins. Subsequent iterations produce more globally optimal model structures than the previous [@guo2015a]. In `NonSysId', $\text{iOFR}_{S}$ extends the iOFR with simulation-based stability tests (bounded-input bounded-output) and model selection. The candidate models generated at each iteration of iOFR are simulated using a constant input of zeros and ones separately. If the corresponding outputs of a model remain bounded with a small variance, typically less than $10^{-2}$, it is regarded as stable. From the stable set of models, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004], based on the mean squared simulated error (MSSE) [@Piroddi2003], selects the best model $\overline{m}$. 
-
-Let $\mathcal{D}$ include both candidate linear and nonlinear terms (monomials) for NARX model identification. As the polynomial nonlinearity, $N_p$, and the number of past outputs and inputs considered increase, $|\mathcal{D}|$ can grow exponentially [@chen1989b; @billings1987a], increasing computational cost for iOFR/$\text{iOFR}_S$. The `NonSysId` package implements four strategies for reducing computational time (RCT) based on what is proposed in [@Wei2004] and [@guo2015a] to speed up convergence toward an optimum model. 
[…]
