### 2025-07-17 a5b74ef
Update paper.md
