### 2025-07-17 435f467
Update paper.md
