### 2025-07-17 a72bbc1
Update paper.md
