+++ paper/paper.md
@@ -30,7 +30,7 @@ System identification involves constructing mathematical models of complex dynam
 
 # Statement of Need
 
-System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. In this setting, system identification determines a functional mapping between past input/output terms (input/output lagged terms) to the current output. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models, which are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the NARMAX model. Additionally, the NOFRF concept [@Lang2005] facilitates a detailed frequency-domain understanding of how nonlinearities affect input-output dynamics captured by NARX models [@Lang2005;@BAYMA2018;@Gunawardena2018;@ZHU2022]. This enhances the utility of NARX models by offering a comprehensive framework for analysing and interpreting nonlinear systems [@Zhu2021].
+System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. In this setting, system identification determines a functional mapping between past input/output terms (input/output lagged terms) to the current output. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models, which are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the NARMAX model.
 
[…]
