### 2025-07-17 b50ab43
Update paper.md
