+++ paper/paper.md
@@ -30,7 +30,7 @@ System identification involves constructing mathematical models of complex dynam
 
 # Statement of Need
 
-System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the widely used discrete-time nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. In this setting, system identification determines a functional mapping between past input/output terms (input/output lagged terms) to the current output. This mapping may be polynomial, a neural network, or even a fuzzy logic-based model. We focus on polynomial NARX models, which are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the NARMAX model.
+System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the discrete-time polynomial nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. Here, system identification determines a polynomial functional mapping between past input/output terms (input/output lagged terms) to the current output. Polynomial NARX models are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the NARMAX model.
 
 The key challenge in learning a polynomial NARX model is to identify a parsimonious model structure from a large set of candidate terms (monomials), as inclusion of superfluous terms in the model can lead to spurious dynamics [@AGUIRRE1995;@mendes1998a]. The Orthogonal Forward Regression (OFR) algorithm, also known as Forward Regression Orthogonal Least Squares (FROLS) [@chen1989b; @billings1987a], can evaluate the independent contribution of each candidate term to the output according to a term selection criterion [@korenberg1988a; @WANG1996; @hong2003], enabling efficient step-wise forward selection of terms. Using the error reduction ratio (ERR), a popular term selection criterion, the OFR incrementally chooses terms that maximise the explained variance (model fit). Recently, two open-sourced packages for system identification have been introduced: `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive. However, they are based on the original FROLS/OFR algorithm. Concerns persist regarding the original OFR algorithm (OFR-ERR), which relies on the ERR for term selection [@Piroddi2003;@Mao1997]. These concerns are mainly regarding the selection of redundant/incorrect terms, overfitting and inaccuracies/instability in long-horizon prediction (simulation, free-run or model-predicted output). These are due to inputs which do not persistently excite the system under consideration (i.e., input not informative to effectively stimulate the system) or the presence of complex noise structures [@Piroddi2003;@Mao1997]. 
 
