### 2025-07-17 b94d578
Update paper.md
