### 2025-07-18 95e9724
Update paper.md
