### 2025-07-21 cfe5f07
Update paper.md
