### 2025-07-21 ec4a0e7
Update paper.md
