### 2025-07-21 7b0db6c
Update paper.md
