### 2025-07-21 ae99402
Update paper.md
