### 2025-07-21 d965edb
Update paper.md
