+++ paper/paper.md
@@ -32,15 +32,13 @@ System identification involves constructing mathematical models of complex dynam
 
 System identification is at the intersection of control theory, dynamic systems theory and machine learning, which aims to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. Its objectives are (i) accurate input–output mapping for prediction and (ii) faithful capture of underlying dynamics of the actual system [@ljung1998system; @billings2013a]. This paper focuses on the discrete-time polynomial nonlinear auto-regressive with exogenous input (NARX) models, with ARX as the linear special case. Here, the polynomial functional mapping is between past input/output terms (input/output lagged terms) to the current output. Polynomial NARX models are equivalent to RNNs [@Sum1999] and widely used in fields such as control, condition monitoring, advanced manufacturing, and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Extending the NARX model with a noise model yields the NARMAX model.
 
-A core challenge is identifying a parsimonious structure from a large set of candidate terms, as redundant terms can induce spurious dynamics [@AGUIRRE1995; @mendes1998a]. The Orthogonal Forward Regression (OFR) algorithm, also known as FROLS [@chen1989b; @billings1987a], can evaluate the independent contribution of each candidate term to the output according to a term selection criterion [@korenberg1988a; @WANG1996; @hong2003], enabling efficient step-wise forward selection of terms. Using the error reduction ratio (ERR), a common criterion, the OFR incrementally chooses terms that maximise the explained variance (model fit). Recently, two open-sourced packages for system identification have been introduced: `SysIdentPy` [@Lacerda2020] for Python and the `narmax` package [@AYALA2020] for R. Both packages are well-developed and comprehensive and are based on the original OFR algorithm. However, the traditional OFR method (using ERR) has known concerns [@Piroddi2003; @Mao1997]. These include the selection of redundant/incorrect terms, overfitting and inaccuracies/instability in long-horizon prediction (simulation, free-run or model-predicted output). These are due to inputs that do not persistently excite the system under consideration (i.e., not informative enough to stimulate the system effectively) or the presence of complex noise structures [@Piroddi2003;@Mao1997]. 
-
-Beyond obtaining parsimony, models must generalise well to unseen data--assessed through suitable cross-validation strategies [@Little2017; @Stone1974]. However, in some real-time applications, such as structural health monitoring or fault diagnosis, obtaining separate validation data may be impractical [@Gharehbaghi2022; @Vamsikrishna2024]. In neuroscience, rapidly changing complex dynamics between brain regions make it difficult to obtain data which precisely contains the same behaviour for validation purposes [@Kunjan2021; @Seedat2024; @Chen2016; @Eichenbaum2021; @Lehnertz2021]. The following section outlines the features in the `NonSysId` package, previously proposed to address these challenges when applying system identification.
[…]
