### 2025-07-21 3d7f02c
Update paper.md
