### 2025-07-21 c6159ea
Update paper.md
