+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-System identification is at the intersection of control theory, dynamic systems theory and machine learning, which involves constructing mathematical models of complex dynamic linear or nonlinear systems using input-output data, enabling the prediction of system behaviour analysis in both time and frequency domains. This approach can model the entire system or capture specific dynamics within it. For meaningful analysis, it is essential for the model to accurately reflect the underlying system behaviour. This paper introduces NonSysId, an open-source MATLAB software package designed for nonlinear system identification, specifically focusing on NARMAX models. The software incorporates an advanced term selection methodology that prioritises simulation (free-run) accuracy while preserving model parsimony. A key feature is the integration of iterative Orthogonal Forward Regression (iOFR) with Predicted Residual Sum of Squares (PRESS) statistic-based term selection, facilitating robust model generalisation without the need for a separate validation dataset. Furthermore, techniques for reducing computational overheads are implemented. These features make `NonSysId` particularly suitable for real-time applications, such as structural health monitoring, fault diagnosis, and biomedical signal processing, where it is challenging to capture signals under consistent conditions, resulting in limited or no validation data.
+System identification is at the intersection of control theory, dynamic systems theory and machine learning, which involves constructing mathematical models of complex dynamic linear or nonlinear systems using input-output data, enabling the prediction of system behaviour and analysis in both time and frequency domains. This approach can model the entire system or capture specific dynamics within it. For meaningful analysis, it is essential for the model to accurately reflect the underlying system behaviour. This paper introduces NonSysId, an open-source MATLAB software package designed for nonlinear system identification, specifically focusing on NARMAX models. The software incorporates an advanced term selection methodology that prioritises simulation (free-run) accuracy while preserving model parsimony. A key feature is the integration of iterative Orthogonal Forward Regression (iOFR) with Predicted Residual Sum of Squares (PRESS) statistic-based term selection, facilitating robust model generalisation without the need for a separate validation dataset. Furthermore, techniques for reducing computational overheads are implemented. These features make `NonSysId` particularly suitable for real-time applications, such as structural health monitoring, fault diagnosis, and biomedical signal processing, where it is challenging to capture signals under consistent conditions, resulting in limited or no validation data.
 
 # Statement of Need
 
