### 2025-07-22 d9dc7e9
Update paper.md
