### 2025-07-22 8e18674
Update paper.md
