### 2025-07-22 297aae1
Update paper.md
