### 2025-07-30 cd4c4f3
Update paper.md
