### 2025-07-30 8f3d869
Update paper.md
