### 2025-07-30 6d63def
Update paper.md
