### 2025-07-30 fe9be55
Update paper.md
