### 2025-07-30 22263c5
Update paper.md
