### 2025-07-30 360eb39
Update paper.md
