### 2025-07-30 e2b8e90
Update paper.bib
