+++ paper/paper.bib
@@ -636,3 +636,11 @@ issn = {0278-6125},
 doi = {https://doi.org/10.1016/j.jmsy.2024.06.004},
 author = {Zepeng Liu and Zi-Qiang Lang and Yufei Gui and Yun-Peng Zhu and Hatim Laalej}
 }
+
+@misc{GitHub_NonSysID,
+author = {Gunawardena Rajintha and Zi-Qiang Lang and Fei He},
+title = {{G}it{H}ub - raj-gun/{N}on{S}ys{I}{D}: {A} {M}at{L}ab package for {S}ystem {I}dentification using linear and nonlinear auto-regresive models ({N}){A}{R}, ({N}){A}{R}{X} and ({N}){A}{R}{M}{A}{X} models --- github.com},
+howpublished = {\url{https://github.com/raj-gun/NonSysID}},
+year = {},
+note = {[Accessed 30-07-2025]},
+}
