### 2025-07-30 874bb5d
Update paper.md
