+++ paper/paper.md
@@ -38,7 +38,7 @@ The core challenge is identifying a parsimonious structure from a large set of c
 
 The `NonSysId` package introduced in this paper implements an OFR-based system identification methodology designed to address the key issues outlined in the previous section. This is achieved by integrating and extending several OFR variants already available in the literature [@guo2015a;@WANG1996;@hong2003], and a simulation-based model selection procedure proposed here. At the time of writing, `NonSysId` is the only open-source package that implements previously proposed solutions to overcome the limitations of the original OFR algorithm. 
 
-To address these limitations, the `NonSysId` package incorporates the iterative OFR (iOFR) algorithm [@guo2015a], simulation-based stability tests (bounded-input bounded-output), and model selection techniques for improved long-horizon prediction and stability. Specifically, the candidate models generated at each iteration of iOFR are simulated using constant inputs of zeros and ones separately. If the corresponding model outputs remain bounded with a small variance, typically less than $10^{-2}$, the model is regarded as stable. From this set of stable models, the Bayesian Information Criterion (BIC) [@Schwarz1978; @Stoica2004], based on the mean squared simulated error (MSSE) [@Piroddi2003], is used to select the best model for the current iteration. Furthermore, the package integrates the PRESS-statistic-based [@Allen1974] term selection criterion [@WANG1996; @hong2003], enabling efficient stepwise forward selection of terms that minimise the one-step-ahead leave-one-out cross-validation error. These enhancements provide robust term selection (compared to OFR-ERR), built-in cross-validation, and models with improved long-horizon prediction capabilities and simulation stability [@AGUIRRE2010]. For NARX models, the candidate term set can be extensive and computationally demanding in the iFRO algorithm [@chen1989b; @billings1987a; @guo2015a]. `NonSysId`, following [@Wei2004] and [@guo2015a], incorporates strategies to accelerate the iterative stepwise forward selection process. Additionally, the package includes correlation-based residual analysis techniques for validating nonlinear models [@Billings1983]. These features make `NonSysId` particularly suitable for real-time applications where inputs may not be persistently exciting and separate validation datasets may be unavailable.
[…]
