### 2025-07-30 a380f35
Update paper.bib
