### 2025-07-30 c4bb4f1
Update paper.md
