+++ paper/supplimentray_information/README.md
@@ -1,1 +1,248 @@
+---
+title: 'NonSysId: Nonlinear System Identification with Improved Model Term Selection for NARMAX Models'
+tags:
+  - System Identification
+  - NARX
+  - NARMAX
+  - Orthogonal Least Squares
+  - Time-series
+  - Nonlinear dynamics
+authors:
+  - name: Rajintha Gunawardena
+    orcid: 0000-0003-1192-9125
+    affiliation: 1
+  - name: Zi-Qiang Lang
+    affiliation: 2
+  - name: Fei He
+    orcid: 0000-0003-1192-9125
+    affiliation: 1
+affiliations:
+  - index: 1
+    name: Centre for Computational Science and Mathematical Modelling, Coventry University, Coventry CV1 5FB, UK
+  - index: 2
+    name: School of Electrical and Electronic Engineering, The University of Sheffield, Western Bank, Sheffield S10 2TN, UK
+date: 26 November 2024
+bibliography: paper.bib
+---
+
+# Summary
+System identification involves constructing mathematical models of dynamic systems using input-output data, enabling analysis and prediction of system behaviour in both time and frequency domains. This approach can model the entire system or/and capture specific dynamics within it. For meaningful analysis, it is essential for the model to accurately reflect the underlying system's behaviour. This paper introduces NonSysId, an open-sourced MATLAB software package designed for nonlinear system identification, specifically focusing on NARMAX models. The software incorporates an advanced term selection methodology that prioritises on simulation (free-run) accuracy while preserving model parsimony. A key feature is the integration of iterative Orthogonal Forward Regression (iOFR) with Predicted Residual Sum of Squares (PRESS) statistic-based term selection, facilitating robust model generalisation without the need for a separate validation dataset. Furthermore, techniques for reducing computational overheads are implemented. These features make NonSysId particularly suitable for real-time applications such as structural health monitoring, fault diagnosis, and biomedical signal processing, where it is a challenge to capture the signals under consistent conditions, resulting in limited or no validation data.
+
+# Statement of Need
+
+System identification is a field at the intersection of control theory, dynamic systems theory and machine learning that seeks to derive mathematical models of dynamic linear or nonlinear systems based on experimental input-output data. 
+Generally, system identification has two primary objectives [@ljung1998system;@billings2013a], (i) to accurately map the system's inputs and outputs, allowing for the prediction of new, unseen data, and (ii) to capture the underlying dynamics of the system within the model.
+
+The dynamic models generated through system identification can be either discrete or continuous time models [@UNBEHAUEN1997]. This paper centres on widely-used discrete-time models, specifically nonlinear auto-regressive models with exogenous inputs (NARX), where the ARX model is a linear variant of the NARX framework. These input-output time-series models predict future outputs of a system based on its historical input and output instances. NARX models have been applied extensively to model and analyse complex systems in fields such as control, condition monitoring, advanced manufacturing and the modelling and analysis of physiological and biological systems [@Chiras2002;@WANG2024;@ZAINOL2022;@RITZBERGER2017;@Gao2023;@HE2016;@HE2021;@LIU2024]. Moreover, it has been demonstrated that the NARX model has equivalence to a recurrent neural network (RNN) [@Sum1999]. Extending the NARX model to incorporate a noise model, we obtain the nonlinear auto-regressive moving average with exogenous inputs (NARMAX) model.  
+
[…]
