+++ paper/supplimentray_information/README.md
@@ -39,51 +39,51 @@ Recently two open-sourced packages have been introduced, `SysIdentPy` [@Lacerda2
 
 In the context of (N)ARX models, system identification is employed to determine a specific functional relationship that maps past input instances (input-lagged terms),
 
-$$
+\begin{equation}\label{eq:Ut_sysid}
     U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
-$$
+\end{equation}
 
 and past output instances (output-lagged terms),
 
-$$
+\begin{equation}\label{eq:Yt_sysid}
     Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\}, 
-$$
+\end{equation}
 
-to the present output instance in time \(y(t)\). \(t\) here refers to a time index (i.e. \(t\)^th^ sample). \(n_a\) and \(n_b\) are the maximum number of past output and input time instances considered and are related to the Lyapunov exponents of the actual system that is being modelled [@mendes1998a]. The functional mapping is described by the following equation:
+to the present output instance in time $y(t)$. $t$ here refers to a time index (i.e. $t$^th^ sample). $n_a$ and $n_b$ are the maximum number of past output and input time instances considered and are related to the Lyapunov exponents of the actual system that is being modelled [@mendes1998a]. The functional mapping is described by the following equation:
 
-$$
+\begin{equation}\label{eq:sys_id_func}
 y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),
-$$
+\end{equation}
 
-where \(y(t)\) and \(u(t)\) refer to the output and input respectively, while \(\xi(t)\) represents the error between the predicted output \(f^{P}\bigl( Y, U \bigr)\) and the actual output \(y(t)\) at time instance \(t\). \(\xi(t)\) will contain noise and unmodeled dynamics. \(f^{P}( \ )\) is the functional mapping between the past inputs and outputs to the current output \(y(t)\). This mapping can take the form of a polynomial, a neural network, or even a fuzzy logic-based model. Here, we focus on polynomial NARX models with a maximum polynomial degree \(N_p \in \mathbb{Z}^{+}\). In this case, \autoref{eq:sys_id_func} can be expressed as
+where $y(t)$ and $u(t)$ refer to the output and input respectively, while $\xi(t)$ represents the error between the predicted output $f^{P}\bigl( Y, U \bigr)$ and the actual output $y(t)$ at time instance $t$. $\xi(t)$ will contain noise and unmodeled dynamics. $f^{P}( \ )$ is the functional mapping between the past inputs and outputs to the current output $y(t)$. This mapping can take the form of a polynomial, a neural network, or even a fuzzy logic-based model. Here, we focus on polynomial NARX models with a maximum polynomial degree $N_p \in \mathbb{Z}^{+}$. In this case, \autoref{eq:sys_id_func} can be expressed as
 
-$$
+\begin{equation}\label{eq:sys_id_func_summation}
 y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),
-$$
+\end{equation}
 
-where \(m = 1, \cdots, M\), \(M\) being the total number of variables or model terms. \(\theta_{m}\) are the model parameters or coefficients and \(\phi_{m}(t)\) are the corresponding model terms or variables. \(\phi_{m}(t)\) are \(n\)^th^-order monomials of the polynomial NARX model \(f^{P}( \ )\), where \(n = 1, \cdots, N_p\) is the degree of the monomial. \(\phi_{m}(t)\) is composed of past output and input time instances from \(Y\) and \(U\). An example of a polynomial NARX model can be
[…]
