### 2025-08-04 6e07be3
Delete paper/supplimentray_information/README.md
