### 2025-08-04 3b7107c
Update supp_info.md
