### 2025-08-04 a014f17
Update supp_info.md
