### 2025-08-04 ce97cce
Update supp_info.md
