### 2025-08-04 bb76fef
Update supp_info.md
