### 2025-08-04 b17b75d
Update supp_info.md
