### 2025-08-04 af8225a
Update supp_info.md
