+++ supplimentray_information/supp_info.md
@@ -1,288 +1,4 @@
 ---
-author:
-- Rajintha Gunawardena<sup>1</sup>
-- Zi-Qiang Lang<sup>2</sup>
-- Fei He<sup>1</sup>
-autoEqnLabels: false
-autoSectionLabels: false
-bibliography: citations.bib
-ccsDelim: ", "
-ccsLabelSep:  —
-ccsTemplate: |
-  ``` math
-  i
-  ```
-  ``` math
-  ccsLabelSep
-  ```
-  ``` math
-  t
-  ```
-chapDelim: .
-chapters: false
-chaptersDepth: 1
-codeBlockCaptions: false
-cref: false
-crossrefYaml: pandoc-crossref.yaml
-csl: ieee.csl
-eqLabels: arabic
-eqnBlockInlineMath: false
-eqnBlockTemplate: |
-  <table>
-  <colgroup>
-  <col style="width: 90%" />
-  <col style="width: 10%" />
-  </colgroup>
-  <tbody>
-  <tr>
-  <td style="text-align: center;"><span
-  class="math display"><em>t</em></span></td>
-  <td style="text-align: right;"><span
-  class="math display"><em>i</em></span></td>
-  </tr>
-  </tbody>
-  </table>
-eqnDisplayTemplate: |
-  ``` math
-  e
-  ```
-eqnIndexTemplate: |
-  (
-  ``` math
-  i
-  ```
-  )
-eqnInlineTableTemplate: |
-  ``` math
-  e
-  ```
-eqnInlineTemplate: |
-  ``` math
-  e
-  ```
-  ``` math
-  equationNumberTeX
-  ```
-  {
-  ``` math
-  i
-  ```
-  }
-eqnPrefix:
-- eq.
-- eqns.
-eqnPrefixTemplate: |
-  ``` math
-  p
-  ```
-   
-  ``` math
-  i
-  ```
-equationNumberTeX: \qquad
-figLabels: arabic
-figPrefix:
-- fig.
-- figs.
-figPrefixTemplate: |
-  ``` math
-  p
-  ```
-   
-  ``` math
-  i
-  ```
-figureTemplate: |
-  ``` math
-  figureTitle
-  ```
-  ``` math
-  i
-  ```
-  ``` math
-  titleDelim
-  ```
-  ``` math
-  t
-  ```
-figureTitle: Figure
-lastDelim: ", "
-linkReferences: false
-listings: false
-listingTemplate: |
-  ``` math
-  listingTitle
-  ```
-  ``` math
-  i
-  ```
-  ``` math
-  titleDelim
-  ```
-  ``` math
-  t
-  ```
-listingTitle: Listing
-listItemTitleDelim: .
-listOfMetadata: false
-lofItemTemplate: |
-  ``` math
-  lofItemTitle
-  ```
-  ``` math
-  i
-  ```
-  ``` math
-  listItemTitleDelim
-  ```
-  ``` math
-  t
-  ```
-    
-lofTitle: |
-  # List of Figures
-lolItemTemplate: |
-  ``` math
-  lolItemTitle
-  ```
-  ``` math
-  i
-  ```
-  ``` math
-  listItemTitleDelim
-  ```
-  ``` math
-  t
-  ```
-    
-lolTitle: |
-  # List of Listings
-lotItemTemplate: |
-  ``` math
-  lotItemTitle
-  ```
-  ``` math
-  i
-  ```
-  ``` math
-  listItemTitleDelim
-  ```
-  ``` math
-  t
-  ```
-    
-lotTitle: |
-  # List of Tables
-lstLabels: arabic
-lstPrefix:
-- lst.
-- lsts.
-lstPrefixTemplate: |
[…]
