### 2025-08-04 a027908
Update supp_info.md
