### 2025-08-04 e6fb4df
Update supp_info.md
