### 2025-08-04 601691b
Update supp_info.md
