+++ supplimentray_information/supp_info.md
@@ -1,7 +1,317 @@
-# NonSysId: Nonlinear System Identification with Improved Model Term Selection for NARMAX Models  
-*An open-source MATLAB package for system identification of ARX, NARX and (N)ARMAX models, featuring improved term selection and robust long-term simulation capabilities.*
+---
+abstract: |
+  # Summary
 
-Authors: Rajintha Gunawardena (https://github.com/raj-gun), Zi-Qiang Lang, Fei He (https://github.com/feihelab)
+  System identification involves constructing mathematical models of
+  dynamic systems using input-output data, enabling analysis and
+  prediction of system behaviour in both time and frequency domains.
+  This approach can model the entire system or/and capture specific
+  dynamics within it. For meaningful analysis, it is essential for the
+  model to accurately reflect the underlying system’s behaviour. This
+  paper introduces NonSysId, an open-sourced MATLAB software package
+  designed for nonlinear system identification, specifically focusing on
+  NARMAX models. The software incorporates an advanced term selection
+  methodology that prioritises on simulation (free-run) accuracy while
+  preserving model parsimony. A key feature is the integration of
+  iterative Orthogonal Forward Regression (iOFR) with Predicted Residual
+  Sum of Squares (PRESS) statistic-based term selection, facilitating
+  robust model generalisation without the need for a separate validation
+  dataset. Furthermore, techniques for reducing computational overheads
+  are implemented. These features make NonSysId particularly suitable
+  for real-time applications such as structural health monitoring, fault
+  diagnosis, and biomedical signal processing, where it is a challenge
+  to capture the signals under consistent conditions, resulting in
+  limited or no validation data.
+
+  **URL:** [https://github.com/raj-gun/NonSysID](URL)
+author:
+- Rajintha Gunawardena<sup>1</sup>
+- Zi-Qiang Lang<sup>2</sup>
+- Fei He<sup>1</sup>
+autoEqnLabels: false
+autoSectionLabels: false
+bibliography: citations.bib
+ccsDelim: ", "
+ccsLabelSep:  —
+ccsTemplate: |
+  ``` math
+  i
+  ```
+  ``` math
+  ccsLabelSep
+  ```
+  ``` math
+  t
+  ```
+chapDelim: .
+chapters: false
+chaptersDepth: 1
+codeBlockCaptions: false
+cref: false
+crossrefYaml: pandoc-crossref.yaml
+csl: ieee.csl
+eqLabels: arabic
+eqnBlockInlineMath: false
+eqnBlockTemplate: |
+  <table>
+  <colgroup>
+  <col style="width: 90%" />
+  <col style="width: 10%" />
+  </colgroup>
+  <tbody>
+  <tr>
+  <td style="text-align: center;"><span
+  class="math display"><em>t</em></span></td>
+  <td style="text-align: right;"><span
+  class="math display"><em>i</em></span></td>
+  </tr>
+  </tbody>
+  </table>
+eqnDisplayTemplate: |
+  ``` math
[…]
