### 2025-08-04 a237012
Update supp_info.md
