+++ supplimentray_information/supp_info.md
@@ -355,8 +355,8 @@ selection process to address these challenges.
 In the context of (N)ARX models, system identification is employed to
 determine a specific functional relationship that maps past input
 instances (input-lagged terms),
-``` math
 {#eq:Ut_sysid}
+``` math
     U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
 ```
 and past output instances (output-lagged terms),
