+++ supplimentray_information/supp_info.md
@@ -355,14 +355,14 @@ selection process to address these challenges.
 In the context of (N)ARX models, system identification is employed to
 determine a specific functional relationship that maps past input
 instances (input-lagged terms),
-{#eq:Ut_sysid}
 ``` math
-    U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
+U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
+    \label{eq:Ut_sysid}
 ```
 and past output instances (output-lagged terms),
 ``` math
-{#eq:Yt_sysid}
-    Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\},
+Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\}, 
+    \label{eq:Yt_sysid}
 ```
 to the present output instance in time $`y(t)`$. $`t`$ here refers to a
 time index (i.e. $`t`$<sup>th</sup> sample). $`n_a`$ and $`n_b`$ are the
@@ -371,8 +371,8 @@ are related to the Lyapunov exponents of the actual system that is being
 modelled \[17\]. The functional mapping is described by the following
 equation:
 ``` math
-{#eq:sys_id_func}
 y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),
+\label{eq:sys_id_func}
 ```
 where $`y(t)`$ and $`u(t)`$ refer to the output and input respectively,
 while $`\xi(t)`$ represents the error between the predicted output
@@ -386,8 +386,8 @@ degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq.
 <a href="#eq:sys_id_func" data-reference-type="eqref"
 data-reference="eq:sys_id_func">[eq:sys_id_func]</a> can be expressed as
 ``` math
-{#eq:sys_id_func_summation}
 y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),
+\label{eq:sys_id_func_summation}
 ```
 where $`m = 1, \cdots, M`$, $`M`$ being the total number of variables or
 model terms. $`\theta_{m}`$ are the model parameters or coefficients and
@@ -398,8 +398,8 @@ the monomial. $`\phi_{m}(t)`$ is composed of past output and input time
 instances from $`Y`$ and $`U`$. An example of a polynomial NARX model
 can be
 ``` math
-{#eq:narx_exmpl}
-    y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
+y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
+    \label{eq:narx_exmpl}
 ```
 In this example, $`\phi_{1}(t)=y(t-1)`$ and $`\phi_{2}(t)=u(t-2)`$ have
 a degree of 1 and are the linear terms (1<sup>st</sup> order monomials
@@ -415,8 +415,8 @@ Eq. <a href="#eq:sys_id_func_summation" data-reference-type="ref"
 data-reference="eq:sys_id_func_summation">[eq:sys_id_func_summation]</a>
 can be represented in matrix form as
 ``` math
-{#eq:sys_id_func_mat}
[…]
