### 2025-08-04 e571e3b
Update supp_info.md
