### 2025-08-04 b3782e7
Update supp_info.md
