+++ supplimentray_information/supp_info.md
@@ -357,11 +357,13 @@ determine a specific functional relationship that maps past input
 instances (input-lagged terms),
 ``` math
 U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
-```{#eq:Ut_sysid}
+```
+{#eq:Ut_sysid}
 and past output instances (output-lagged terms),
 ``` math
 Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\}, 
-```{#eq:Yt_sysid}
+```
+{#eq:Yt_sysid}
 to the present output instance in time $`y(t)`$. $`t`$ here refers to a
 time index (i.e. $`t`$<sup>th</sup> sample). $`n_a`$ and $`n_b`$ are the
 maximum number of past output and input time instances considered and
@@ -370,7 +372,8 @@ modelled \[17\]. The functional mapping is described by the following
 equation:
 ``` math
 y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),
-```{#eq:sys_id_func}
+```
+{#eq:sys_id_func}
 where $`y(t)`$ and $`u(t)`$ refer to the output and input respectively,
 while $`\xi(t)`$ represents the error between the predicted output
 $`f^{P}\bigl( Y, U \bigr)`$ and the actual output $`y(t)`$ at time
@@ -384,7 +387,8 @@ degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq.
 data-reference="eq:sys_id_func">[eq:sys_id_func]</a> can be expressed as
 ``` math
 y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),
-```{#eq:sys_id_func_summation}
+```
+{#eq:sys_id_func_summation}
 where $`m = 1, \cdots, M`$, $`M`$ being the total number of variables or
 model terms. $`\theta_{m}`$ are the model parameters or coefficients and
 $`\phi_{m}(t)`$ are the corresponding model terms or variables.
@@ -395,7 +399,8 @@ instances from $`Y`$ and $`U`$. An example of a polynomial NARX model
 can be
 ``` math
 y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
-```{#eq:narx_exmpl}
+```
+{#eq:narx_exmpl}
 In this example, $`\phi_{1}(t)=y(t-1)`$ and $`\phi_{2}(t)=u(t-2)`$ have
 a degree of 1 and are the linear terms (1<sup>st</sup> order monomials
 or linear monomials) of the model.
@@ -411,7 +416,8 @@ data-reference="eq:sys_id_func_summation">[eq:sys_id_func_summation]</a>
 can be represented in matrix form as
 ``` math
 \mathbf{Y} = \mathbf{\Phi} \mathbf{\Theta} + \mathbf{\Xi},
-```{#eq:sys_id_func_mat}
+```
+{#eq:sys_id_func_mat}
 where $`\mathbf{Y} = \left[ y(1), \cdots, y(L) \right]^T`$ is the vector
 containing the output samples $`y(t)`$.
 $`\mathbf{\Phi} = \left[ \bar{\phi}_{1}, \cdots, \bar{\phi}_{M} \right]`$,
@@ -443,7 +449,8 @@ candidate terms could be
               y(t-1)u(t-2), y(t-2)u(t-1)^{3}, 
[…]
