### 2025-08-04 e6dd744
Update supp_info.md
