### 2025-08-04 f686c5d
Update supp_info.md
