### 2025-08-04 e5f76b5
Update supp_info.md
