+++ supplimentray_information/supp_info.md
@@ -49,7 +49,7 @@ instances (input-lagged terms),
 ``` math
 U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},
 ```
-(1)
+                                                        (1)
 {#eq:Ut_sysid}
 
 and past output instances (output-lagged terms),
