### 2025-08-04 4a4a945
Update supp_info.md
