### 2025-08-04 a73bd13
Update supp_info.md
