### 2025-08-04 7600e87
Update supp_info.md
