### 2025-08-04 58e5c4b
Update supp_info.md
