+++ supplimentray_information/supp_info.md
@@ -53,7 +53,7 @@ U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},   \ \ \ \ \ \ \ \ \ \
 
 and past output instances (output-lagged terms),
 ``` math
-Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\}, 
+Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\},    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (2) 
 ```
 {#eq:Yt_sysid}
 to the present output instance in time $`y(t)`$. $`t`$ here refers to a
@@ -63,7 +63,7 @@ are related to the Lyapunov exponents of the actual system that is being
 modelled \[17\]. The functional mapping is described by the following
 equation:
 ``` math
-y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),
+y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),   \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (3)
 ```
 {#eq:sys_id_func}
 where $`y(t)`$ and $`u(t)`$ refer to the output and input respectively,
@@ -78,7 +78,7 @@ degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq.
 <a href="#eq:sys_id_func" data-reference-type="eqref"
 data-reference="eq:sys_id_func">[eq:sys_id_func]</a> can be expressed as
 ``` math
-y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),
+y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),   \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (4)
 ```
 {#eq:sys_id_func_summation}
 where $`m = 1, \cdots, M`$, $`M`$ being the total number of variables or
@@ -90,7 +90,7 @@ the monomial. $`\phi_{m}(t)`$ is composed of past output and input time
 instances from $`Y`$ and $`U`$. An example of a polynomial NARX model
 can be
 ``` math
-y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).
+y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (5)
 ```
 {#eq:narx_exmpl}
 In this example, $`\phi_{1}(t)=y(t-1)`$ and $`\phi_{2}(t)=u(t-2)`$ have
@@ -107,7 +107,7 @@ Eq. <a href="#eq:sys_id_func_summation" data-reference-type="ref"
 data-reference="eq:sys_id_func_summation">[eq:sys_id_func_summation]</a>
 can be represented in matrix form as
 ``` math
-\mathbf{Y} = \mathbf{\Phi} \mathbf{\Theta} + \mathbf{\Xi},
+\mathbf{Y} = \mathbf{\Phi} \mathbf{\Theta} + \mathbf{\Xi},    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (6)
 ```
 {#eq:sys_id_func_mat}
[…]
