+++ supplimentray_information/supp_info.md
@@ -47,7 +47,7 @@ determine a specific functional relationship that maps past input
 instances (input-lagged terms),
 
 ``` math
-U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},   \null\hfill    (1)
+U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},   &&    (1)
 ```
 {#eq:Ut_sysid}
 
