### 2025-08-04 bfa6be2
Update supp_info.md
