### 2025-08-04 6af3274
Update supp_info.md
