### 2025-08-04 0f0ca7f
Update supp_info.md
