### 2025-08-04 3678b53
Update supp_info.md
