### 2025-08-04 c424381
Update supp_info.md
