+++ supplimentray_information/supp_info.md
@@ -49,23 +49,24 @@ instances (input-lagged terms),
 ``` math
 U = \Big\{ u(t-1)\ ,\ u(t-2)\ ,\ \cdots,\ u(t-n_b) \Big\},    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (1)
 ```
-{#eq:Ut_sysid}
 
 and past output instances (output-lagged terms),
+
 ``` math
 Y = \Big\{ y(t-1)\ ,\ y(t-2)\ ,\ \cdots,\ y(t-n_a) \Big\},    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (2) 
 ```
-{#eq:Yt_sysid}
+
 to the present output instance in time $`y(t)`$. $`t`$ here refers to a
 time index (i.e. $`t`$<sup>th</sup> sample). $`n_a`$ and $`n_b`$ are the
 maximum number of past output and input time instances considered and
 are related to the Lyapunov exponents of the actual system that is being
 modelled \[17\]. The functional mapping is described by the following
 equation:
+
 ``` math
 y(t) = f^{P}\bigl( Y, U \bigr) + \xi(t),   \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (3)
 ```
-{#eq:sys_id_func}
+
 where $`y(t)`$ and $`u(t)`$ refer to the output and input respectively,
 while $`\xi(t)`$ represents the error between the predicted output
 $`f^{P}\bigl( Y, U \bigr)`$ and the actual output $`y(t)`$ at time
@@ -74,13 +75,12 @@ $`f^{P}( \ )`$ is the functional mapping between the past inputs and
 outputs to the current output $`y(t)`$. This mapping can take the form
 of a polynomial, a neural network, or even a fuzzy logic-based model.
 Here, we focus on polynomial NARX models with a maximum polynomial
-degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq.
-<a href="#eq:sys_id_func" data-reference-type="eqref"
-data-reference="eq:sys_id_func">[eq:sys_id_func]</a> can be expressed as
+degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq. (3) can be expressed as
+
 ``` math
 y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),   \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (4)
 ```
-{#eq:sys_id_func_summation}
+
 where $`m = 1, \cdots, M`$, $`M`$ being the total number of variables or
 model terms. $`\theta_{m}`$ are the model parameters or coefficients and
 $`\phi_{m}(t)`$ are the corresponding model terms or variables.
@@ -89,27 +89,27 @@ NARX model $`f^{P}( \ )`$, where $`n = 1, \cdots, N_p`$ is the degree of
 the monomial. $`\phi_{m}(t)`$ is composed of past output and input time
 instances from $`Y`$ and $`U`$. An example of a polynomial NARX model
 can be
+
 ``` math
 y(t) = \theta_{1}y(t-1) + \theta_{2}u(t-2) + \theta_{3}y(t-2)^{2}u(t-1)^{3} + \xi(t).    \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \    (5)
[…]
