### 2025-08-04 0de5ed0
Update supp_info.md
