### 2025-08-04 2bac061
Update supp_info.md
