### 2025-08-04 b3c5095
Update supp_info.md
