### 2025-08-04 2fb4c85
Update supp_info.md
