### 2025-08-04 f528792
Update supp_info.md
