### 2025-08-04 4e60808
Update supp_info.md
