### 2025-08-04 a2149a3
Update supp_info.md
