+++ supplimentray_information/README.md
@@ -301,7 +301,6 @@ are as follows:
 <table>
 <tr>
 <td>
-
 1.  Pre-select each term given in $`\mathcal{P}`$ as the first model
     term and search through $`p`$ orthogonalization paths using OFR to
     produce a set of $`p`$ candidate models
@@ -319,7 +318,6 @@ are as follows:
 
 5.  Repeat steps 1-4 and iteratively search through different
     orthogonalization paths.
-
 </td>
 </tr>
 </table>
