+++ supplimentray_information/README.md
@@ -2,7 +2,7 @@
 
 Authors: Rajintha Gunawardena (https://github.com/raj-gun), Zi-Qiang Lang, Fei He (https://github.com/feihelab)
 
-# Statement of need
+# System identification
 
 System identification is a field at the intersection of control theory,
 dynamic systems theory and machine learning that seeks to derive
@@ -10,10 +10,10 @@ mathematical models of dynamic linear or nonlinear systems based on
 experimental input-output data. Generally, system identification has two
 primary objectives \[1\], \[2\], 
 
-(i) to accurately map the system’s
+`(i) to accurately map the system’s
 inputs and outputs, allowing for the prediction of new, unseen data, and
 
-(ii) to capture the underlying dynamics of the system within the model.
+(ii) to capture the underlying dynamics of the system within the model.`
 
 The dynamic models generated through system identification can be either
 discrete or continuous time models \[3\]. This package centres on
@@ -31,7 +31,7 @@ the nonlinear auto-regressive moving average with exogenous inputs
 (NARMAX) model.
 
 Recently, two open-sourced packages have been introduced, `SysIdentPy`
-\[13\] for Python and the ‘narmax’ package \[14\] for R. Both packages
+\[13\] for Python and the `narmax` package \[14\] for R. Both packages
 are well-developed and comprehensive. However, they are based on the
 original forward regression orthogonal least squares (OFR) algorithm,
 which has been noted to have several limitations, as discussed in
