+++ supplimentray_information/README.md
@@ -81,7 +81,7 @@ Here, we focus on polynomial NARX models with a maximum polynomial
 degree $`N_p \in \mathbb{Z}^{+}`$. In this case, Eq. (3) can be expressed as
 
 ``` math
-y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),   \qquad \qquad \qquad \qquad \qquad \qquad \qquad \qquad \qquad   (4)
+y(t) = \sum_{m=1}^{M} \theta_{m} \times \phi_{m}(t) + \xi(t),   \qquad \qquad \qquad \qquad \qquad \qquad \qquad \qquad \ \ \   (4)
 ```
 
 where $`m = 1, \cdots, M`$, $`M`$ being the total number of variables or
