+++ supplimentray_information/README.md
@@ -42,7 +42,7 @@ to excite the actual system sufficiently. Additionally, in some
 applications, acquiring extra data for cross-validation may be
 infeasible. As a result, developing parsimonious models that can
 generalise well to unseen data becomes crucial in such cases. This paper
-introduces the ‘NonSysId’ package, which incorporates an enhanced model
+introduces the `NonSysId` package, which incorporates an enhanced model
 selection process to address these challenges.
 
 In the context of (N)ARX models, system identification is employed to
@@ -212,32 +212,32 @@ within milliseconds. As a result, obtaining electrophysiological data
 that precisely captures such behaviour is often challenging, if not
 impossible \[28\]–\[32\]. These challenges are critical when applying
 system identification to specific domains. The following section
-outlines the features in the ‘NonSysId’ package, designed to address
+outlines the features in the `NonSysId` package, designed to address
 these issues.
 
 # Features in NonSysId
 
-The ‘NonSysId’ package introduced in this paper implements an OFR-based
+The `NonSysId` package introduced in this paper implements an OFR-based
 system identification methodology designed to address the key issues
 mentioned in the latter part of the previous section. This is achieved
 by integrating and extending several OFR variants already available in
 the literature \[22\], \[23\], \[33\], along with a proposed
 simulation-based model selection procedure. A notable feature of
-‘NonSysId’ is the implementation of the iterative-OFR (iFRO) variant
+`NonSysId` is the implementation of the iterative-OFR (iFRO) variant
 \[33\] of the OFR algorithm. Additionally, the PRESS-statistic-based
 term selection \[22\], \[23\] is integrated with the iOFR, complemented
 by simulation-based model selection. These enhancements enable robust
 term selection (compared to the ERR), built-in cross-validation, and the
 ability to produce models with long-horizon prediction capabilities and
-simulation stability \[34\]. With these features, the ‘NonSysId’ package
+simulation stability \[34\]. With these features, the `NonSysId` package
 makes system identification feasible for real-time applications, such as
 fault diagnosis in engineering or the analysis of electrophysiology
 activity in medical settings, where inputs may not be persistently
 exciting and separate datasets for validation may be unavailable.
-‘NonSysId’ is the only open-sourced package that directly address the
+`NonSysId` is the only open-sourced package that directly address the
 limitations of the original OFR algorithm. For NARX models, where the
 candidate term set can be extensive and computationally demanding in the
-iFRO algorithm, ‘NonSysId’ incorporates methods to reduce the candidate
+iFRO algorithm, `NonSysId` incorporates methods to reduce the candidate
 term set, significantly speeding up the forward selection process.
 Moreover, the package includes correlation-based residual analysis
 techniques for nonlinear model validation \[35\].
@@ -295,7 +295,7 @@ iOFR converge faster towards an optimum while improving computational
 efficiency.
 
 In the original iOFR algorithm \[33\], model selection was based on the
-one-step-ahead prediction. The implementation of iOFR in the ‘NonSysId’
+one-step-ahead prediction. The implementation of iOFR in the `NonSysId`
 extends this by incorporating simulation-based model selection
 ($`\text{iOFR}_{S}`$) to ensure simulation stability and improve
 long-horizon prediction accuracy. The procedures for $`\text{iOFR}_{S}`$
@@ -373,7 +373,7 @@ models.
 ## $`\text{iOFR}_{S}`$ with reduced computational time
 
 This section outlines the comprehensive procedures implemented in the
-‘NonSysId’ package for identifying (N)ARX models from system
[…]
