+++ supplimentray_information/README.md
@@ -14,9 +14,9 @@ primary objectives \[1\], \[2\]:
 <tr>
 <td>
   
-   (i) Accurately map the system’s inputs and outputs, allowing for the prediction of new, unseen data, and
+  * Accurately map the system’s inputs and outputs, allowing for the prediction of new, unseen data, and
 
-  (ii) Faithfully capture the underlying dynamics of the system within the model.
+  * Faithfully capture the underlying dynamics of the system within the model.
   
 </td>
 </tr>
