+++ supplimentray_information/README.md
@@ -14,10 +14,10 @@ primary objectives \[1\], \[2\]:
 <tr>
 <td>
 
-    * Accurately map the system’s inputs and outputs, allowing for the prediction of new, unseen data, and
-    
-    * Faithfully capture the underlying dynamics of the system within the model.
-      
+   * Accurately map the system’s inputs and outputs, allowing for the prediction of new, unseen data, and
+
+   * Faithfully capture the underlying dynamics of the system within the model.
+          
 </td>
 </tr>
 </table>
