+++ supplimentray_information/README.md
@@ -153,7 +153,9 @@ candidate terms could be
 
 from which a NARX model structure, such as that in Eq. (5), can be identified.
 Once the model structure is identified, the next step is to estimate the
-model parameters. However, determining the appropriate linear and
+model parameters. 
+
+> However, determining the appropriate linear and
 nonlinear terms to include in the model structure is critical to
 achieving parsimonious models. This is particularly important in the
 nonlinear cases \[2, Ch. 1\], as the inclusion of unnecessary model
