+++ supplimentray_information/README.md
@@ -155,6 +155,7 @@ from which a NARX model structure, such as that in Eq. (5), can be identified.
 Once the model structure is identified, the next step is to estimate the
 model parameters. 
 
+<p>&nbsp;</p>
 
 > **Determining the appropriate linear and
 nonlinear terms to include in the model structure is critical to
@@ -163,6 +164,7 @@ nonlinear cases \[2, Ch. 1\], as the inclusion of unnecessary model
 terms, can result in a model that erroneously captures dynamics that do
 not belong to the underlying system \[17\], \[18\].**
 
+<p>&nbsp;</p>
 
 The Orthogonal Forward Regression (OFR) algorithm, also known as Forward
 Regression OLS (FROLS) \[19\], \[20\], is based on the Orthogonal Least
