+++ supplimentray_information/README.md
@@ -154,7 +154,6 @@ candidate terms could be
 from which a NARX model structure, such as that in Eq. (5), can be identified.
 Once the model structure is identified, the next step is to estimate the
 model parameters. 
-
 <p>&nbsp;</p>
 
 > **Determining the appropriate linear and
@@ -165,7 +164,6 @@ terms, can result in a model that erroneously captures dynamics that do
 not belong to the underlying system \[17\], \[18\].**
 
 <p>&nbsp;</p>
-
 The Orthogonal Forward Regression (OFR) algorithm, also known as Forward
 Regression OLS (FROLS) \[19\], \[20\], is based on the Orthogonal Least
 Squares (OLS). When combined with an appropriate term selection
