+++ supplimentray_information/README.md
@@ -244,20 +244,9 @@ simulation-based model selection procedure.
 <tr>
 <td>
   
-  > **A notable feature of
-  `NonSysId` is the implementation of the `iterative-OFR (iFRO)` variant
-  \[33\] of the OFR algorithm. Additionally, the `PRESS-statistic-based
-  term selection` \[22\], \[23\] is integrated with the iOFR, complemented
-  by `simulation-based model selection`. These enhancements enable robust
-  term selection (compared to the ERR), `built-in cross-validation`, and the
-  ability to `produce models with long-horizon prediction capabilities and
-  simulation stability` \[34\].** 
+> A notable feature of `NonSysId` is the implementation of the ***iterative-OFR (iFRO)*** variant \[33\] of the OFR algorithm. Additionally, the ***PRESS-statistic-based term selection*** \[22\], \[23\] is integrated with the iOFR, complemented by ***simulation-based model selection***. These enhancements enable robust term selection (compared to the ERR), ***built-in cross-validation***, and the ability to ***produce models with long-horizon prediction capabilities and simulation stability*** \[34\]. 
   
-  > **With these features, the `NonSysId` package
-  makes system identification feasible for real-time applications, such as
-  fault diagnosis in engineering or the analysis of electrophysiology
-  activity in medical settings, where inputs may not be persistently
-  exciting and separate datasets for validation may be unavailable.**
+> With these features, the `NonSysId` package makes system identification feasible for real-time applications, such as fault diagnosis in engineering or the analysis of electrophysiology activity in medical settings, where inputs may not be persistently exciting and separate datasets for validation may be unavailable.
 
 </td>
 </tr>
