+++ supplimentray_information/README.md
@@ -310,7 +310,7 @@ $`\mathcal{P}`$ will be discussed in later sections. This will make the
 iOFR converge faster towards an optimum while improving computational
 efficiency.
 
-## $`\text{iOFR}_{S}`$: iOFR with simultion-based model selection
+## iOFR with simultion-based model selection ($`\text{iOFR}_{S}`$)
 
 In the original iOFR algorithm \[33\], model selection was based on the
 one-step-ahead prediction. The implementation of iOFR in the `NonSysId`
