### 2025-08-19 2ab2b97
Update example1a_dc_motor.m
