### 2025-08-19 12c5882
Update example1b_dc_motor.m
