### 2025-08-19 87aaa37
Update electro_mechanical_example.m
