+++ Examples/Electro-mecahnical system/electro_mechanical_example.m
@@ -3,7 +3,8 @@ clear;clc
 addpath('\...\NonSysID');
 %% Import data 
 % The real data in this example is obtained from an electromechanical system described in [1].
-u = readmatrix('\...\Examples\Electro-mecahnical system\Data\x_cc.csv');
+% Replace '\...' below with the path to the Examples folder
+u = readmatrix('\...\Examples\Electro-mecahnical system\Data\x_cc.csv'); 
 y = readmatrix('\...\Examples\Electro-mecahnical system\Data\y_cc.csv');
 
 %--- Down sample data ---%
@@ -62,8 +63,9 @@ tic
     NonSysID(mod_type,u_ID,y_ID,na1,na2,nb1,nb2,nl_ord_max,is_bias,n_inpts,KSA_h,RCT,x_iOFR,stp_cri,D1_thresh,displ,sim,parall);
 toc
 
-disp('ARX model:'); disp(iOFR_table_lin{best_mod_ind_lin,1});
-if best_mod_ind_nl~=0
+% Print ARX/NARX models
+disp('ARX model:'); disp(iOFR_table_lin{best_mod_ind_lin,1}); % Print the best ARX model that fits the data
+if best_mod_ind_nl~=0 % If a NARX model was identified, then display the best NARX model that fits the data
     disp('NARX model:'); 
     tbl_NARX = join(iOFR_table_nl{best_mod_ind_nl,10},iOFR_table_nl{best_mod_ind_nl,1});disp(tbl_NARX);
 end
