+++ Examples/README.md
@@ -1,1 +1,9 @@
+# Example code structure to use the NonSysID package
+This document provides a code structure to do system identification using the NonSysID package. The examples shown here follow the same structure.
 
+In MATLAB, first clear the workspace and add the folder containing the `NonSydID` code to the search path
+
+```
+close all;clear;clc
+addpath('\...\NonSysID');
+```
