+++ Examples/README.md
@@ -136,18 +136,18 @@ ylabel('$y(t)/\hat{y}(t)$','Interpreter','latex','FontSize',12);
 
 ## Validate an identified (N)ARX model using correlation-based validation techniques 
 
-(1) Simulate the identified model ('model'). 'u' is the input and 'y' is the actual output from the real system.
+(1) Simulate the identified model (`model`). `u` is the input and `y` is the actual output from the real system.
 
 ```matlab
 [sse, y_hat, error, U_delay_mat_sim] = model_simulation(model,u,y,KSA_h);
 ```
 
-(2) Specify residual type: simulation, one-step ahead or k-steps ahead (1,2,3 respectively) 
+(2) Specify residual type: `simulation`, `one-step ahead` or `k-steps ahead` (`1`,`2` or `3` respectively) 
 ```matlab
 err_typ = 1;
 ```
 
-(3) Normalise the residuals ('error') and input ('U_delay_mat_sim').
+(3) Normalise the residuals (`error`) and input (`U_delay_mat_sim`, output variable from function `model_simulation` in step (1)).
 ```matlab
 e_ct = error(tt_splt,err_typ); % 'err_typ': simulation, one-step ahead or k-steps ahead (1,2,3 respectively)
 u_ct = U_delay_mat_sim(tt_splt,1);
