+++ Examples/README.md
@@ -159,7 +159,7 @@ e_ct = e_ct ./ std(e_ct);
 
 (4) Correlation-based validation; 
 Here, the variable `max_lag` is the maximum lag used for correlation. This is usually set to the length of the residuals. Set the function variable `plt` to `1` for enabling plots or `0` otherwise.   
-    * For NARX models
++ For NARX models
 ```matlab
 [dat,conf_inv] = ac_cc_model_valid_nl(e_ct,u_ct,max_lag,plt);
 
@@ -167,7 +167,7 @@ disp(['RMSE = ',num2str( sqrt(mean(error(:,1).^2)) )]);
 disp(['Error variance = ',num2str( var(error(:,1)) )]);
 ```
     
-    * For ARX models
++ For ARX models
 ```matlab
 [dat,conf_inv] = ac_cc_model_valid(e_ct, u_ct, max_lag,plt)
 ac_cc_model_valid_nl(e_ct,u_ct,length(tt_splt),1);
