+++ doc/NonSysID.md
@@ -1,0 +1,112 @@
+# `NonSysID`
+
+`NonSysID` performs system identification for **linear/nonlinear ARX** (input–output) and **AR** (autoregressive) models using an iterative Orthogonal Forward Regression (iOFR) algorithm.  
+It supports polynomial nonlinearities, multiple stopping criteria, k-step-ahead prediction, and optional parallelization.
+
+---
+
+## Signature
+
+```matlab
+[model, Mod_Val_dat, iOFR_table_lin, iOFR_table_nl, ...
+ best_mod_ind_lin, best_mod_ind_nl, val_stats] = ...
+    NonSysID(mod_type, u_ID, y_ID, ...
+             na1, na2, nb1, nb2, ...
+             nl_ord_max, is_bias, n_inpts, KSA_h, RCT, ...
+             x_iOFR, stp_cri, D1_thresh, displ, sim, parall);
+```
+---
+
+## Parameters
+
+| Name        | Type        | Required | Description |
+|-------------|-------------|----------|-------------|
+| `mod_type`  | `char`      | Yes 	   | Model type: `'ARX'` (linear/nonlinear input–output model) or `'AR'` (linear/nonlinear autoregressive model). |
+| `u_ID`      | `vector/matrix`    | Yes 	   | Input signal(s) used for identification. Number of columns depends on `n_inpts`. |
+| `y_ID`      | `vector`    | Yes 	   | Output signal used for identification. |
+| `na1`       | `int`       | Yes 	   | Minimum output lag (smallest lagged output term included). |
+| `na2`       | `int`       | Yes 	   | Maximum output lag (largest lagged output term included). |
+| `nb1`       | `int`       | Yes 	   | Minimum input lag (smallest lagged input term included). |
+| `nb2`       | `int`       | Yes 	   | Maximum input lag (largest lagged input term included). |
+| `nl_ord_max`| `int`       | Yes 	   | Maximum polynomial nonlinearity order. `1 = linear`, `>1 = nonlinear`. |
+| `is_bias`   | `int` (0/1) | Yes 	   | Include bias/DC offset term (`0 = include`, `1 = exclude`). |
+| `n_inpts`   | `int`       | Yes 	   | Number of input signals. |
+| `KSA_h`     | `int`       | Yes 	   | Horizon for **k-step-ahead prediction**. |
+| `RCT`       | `int` (0–4) | Yes 	   | Reduce Computational Time method (`0 = none`, `1–4 = selected RCT method`). |
+| `x_iOFR`    | `logical[2]`| Yes 	   | Whether to run more than one iOFR iteration: `x_iOFR(1)` for linear model, `x_iOFR(2)` for nonlinear. |
+| `stp_cri`   | `cell{2}`   | Yes 	   | Stopping criteria: `'PRESS_thresh'` or `'BIC_thresh'`. First for linear, second for nonlinear. |
+| `D1_thresh` | `double[2]` | Yes 	   | Threshold values for stopping criteria. First for linear, second for nonlinear. |
+| `displ`     | `int` (0/1) | Yes 	   | Display all candidate models (`1`) or only final model (`0`). |
+| `sim`       | `int[2]`    | Yes 	   | Run and display simulation: `sim(1)` for linear model, `sim(2)` for nonlinear (`0 = no`, `1 = yes`). |
+| `parall`    | `int[2]`    | Yes 	   | Enable parallel processing: `parall(1)` for linear, `parall(2)` for nonlinear (`0 = no`, `1 = yes`). |
+
+---
+
+## Returns
+
+| Output             | Type     | Description |
+|--------------------|----------|-------------|
[…]
