### 2025-08-20 6526653
Update Example_code_structure.md
