+++ README.md
@@ -42,7 +42,7 @@ Authors: Rajintha Gunawardena (https://github.com/raj-gun), Zi-Qiang Lang, Fei H
     - or use the 'addpath' command in the Matlab script to add the folder 'NonSysId' and use the functions within (https://uk.mathworks.com/help/matlab/ref/addpath.html).
 
 ### Documentation
-Brief documentation explaining the main function and a code structure for identifying a model, simulating and validating an identified model is given in [the `doc` folder](https://github.com/raj-gun/NonSysID/tree/main/doc)
+Brief documentation explaining the main function and a code structure for identifying a model, simulating and validating an identified model is given in the [`doc`](https://github.com/raj-gun/NonSysID/tree/main/doc) folder
 
 ### Examples
 - Basic use of identifying a SISO NARX model from real data, see the example in `Examples/Electro-mecahnical system`.
