### 2025-09-15 a16c332
Update Example_code_structure.md
