### 2025-09-15 00021b7
Addressed issue #3
