+++ Examples/Electro-mecahnical system/electro_mechanical_example.m
@@ -1,118 +1,0 @@
-%% Clear workspace and add NonSydID to the search path
-clear;clc
-addpath('\...\NonSysID');
-%% Import data 
-% The real data in this example is obtained from an electromechanical
-% system described in [1]. The data is available in [2].
-%[1]. W. R. L. Junior, V. M. Almeida, and S. A. Milani, “Identifica¸cao de um motor/gerador cc por meio de modelos polinomiais autorregressivos e redes neurais artificiais,” in Proc. XIII Simp´osio Brasileiro De Automa¸c˜ao Inteligente, pp. 1–6, 2017.
-%[2]. wilsonrljr, “sysidentpy/examples/datasets at main · wilsonrljr/sysidentpy,” GitHub, 2019. https://github.com/wilsonrljr/sysidentpy/tree/main/examples/datasets (accessed Dec. 04, 2024).
-
-% Replace '...' below with the path to the Examples folder
-u = readmatrix('...\Examples\Electro-mecahnical system\Data\x_cc.csv'); 
-y = readmatrix('...\Examples\Electro-mecahnical system\Data\y_cc.csv');
-
-%--- Down sample data ---%
-dwn_smpl = 100;
-u = u(1:dwn_smpl:end);
-y = y(1:dwn_smpl:end);
-%% System Identification - NonSysID
-tt_splt = 100:350; %Training samples of data
-u_ID=u(tt_splt);
-y_ID=y(tt_splt);
-
-% Model type ARX/AR
-mod_type = 'ARX';
-
-% Maximum and minimum output lags
-na1=1;na2=3; 
-
-% Maximum and minimum input lags
-nb1=1;nb2=3; 
-
-% Maximum order of polynomial nonlinearity considered
-nl_ord_max=2; 
-
-% Run more than one iteration of iOFR for [linear model ,nonlinear model]
-x_iOFR = [true,true];
-
-% Stoping criteria for [linear model ,nonlinear model]. PRESS_thresh/BIC_thresh
-stp_cri = {'PRESS_thresh', 'PRESS_thresh'}; 
-% Set value for stopping criteria for [linear model ,nonlinear model]
-% The stopping criteria relates to how many terms are added to the model
-D1_thresh = [10^(-10),10^(0.9)];
-
-% Specify if bias/DC off set is required, 0, or not, 1.
-is_bias=0;
-
-% Specify number of inputs
-n_inpts=1;
-
-% Specify the number of steps for k-steps ahead prediction 
-KSA_h=20;
-
-% Specify which RCT (Reduce Computational Time) method to use, 1-4, 0 for
-% no RCT.
-RCT=4;
-
-% Specify whether to simulate model and display results respectively
-sim=[1,1];
-% Set to 1 to display all models generated from iOFRs, 0 otherwise 
-displ=0; 
-% Set 1 or 0 to use parallel processing to accelerate iOFRs, for [linear model ,nonlinear model]
-parall = [1,1];
-
-% Run NonSysID
-tic
-[model, Mod_Val_dat, iOFR_table_lin, iOFR_table_nl, best_mod_ind_lin, best_mod_ind_nl, val_stats] = ...
-    NonSysID(mod_type,u_ID,y_ID,na1,na2,nb1,nb2,nl_ord_max,is_bias,n_inpts,KSA_h,RCT,x_iOFR,stp_cri,D1_thresh,displ,sim,parall);
-toc
-
-% Print ARX/NARX models
[…]
