### 2025-09-16 10a267e
Update model_simulation function.md
