+++ doc/model_simulation function.md
@@ -1,7 +1,7 @@
 # `model_simulation`
 
 `model_simulation` simulates the output of an identified NARX (Nonlinear AutoRegressive with eXogenous input) model using given input/output data and a specified k-step-ahead horizon.  
-It returns the simulated output, error, and delay matrix, and is typically used after identifying a model with [`NonSysID`](./NonSysID_Documentation.md).
+It returns the simulated output, error, and delay matrix, and is typically used after identifying a model with [`NonSysID`](./NonSysID.md).
 
 ---
 
