### 2025-09-16 7de8bf8
Update model_simulation function.md
