### 2025-09-16 7d61c13
Update model_simulation function.md
