### 2025-09-16 7213d00
Update model_simulation function.md
