### 2025-09-16 4fa3408
Update model_simulation function.md
