### 2025-09-16 6ae652f
Update Example_code_structure.md
