### 2025-09-16 a5d79d9
Update model_simulation function.md
