### 2025-09-16 b5db2e7
Update model_simulation function.md
