### 2025-09-16 7558f17
Update model_simulation function.md
