### 2025-09-16 f28af1b
Update model_simulation function.md
