### 2025-09-16 e141933
Update model_simulation function.md
