+++ doc/NonSysID.md
@@ -62,9 +62,9 @@
 3. **Stopping criteria**: Uses PRESS (Prediction Error Sum of Squares) or BIC (Bayesian Information Criterion), with thresholds in `D1_thresh`.  
 4. **Iteration control**: Optional multiple iOFR iterations until convergence (`x_iOFR`).  
 5. **Performance enhancements**: RCT acceleration and parallelisation options.
-
-For more details, please refer to the supplementary information (NonSysID
-/supplimentray_information/README.md)
+%
+For more details, please refer to the [`supplementary information`](./NonSysID
+/supplimentray_information/README.md).
 ---
 
 ## Example Usage
