+++ doc/NonSysID.md
@@ -113,4 +113,4 @@ parall = [1,1];
 
 ## See Also
 
-- To simulate a NARX model identified using `NonSysID`, refer to [`model_simulation`](./model_simulation function.md)  
+- To simulate a NARX model identified using `NonSysID`, refer to [`model_simulation`](./model_simulation.md)  
