### 2025-09-16 8c1f24c
Update Example_code_structure.md
