### 2025-09-16 c75bac8
Update Example_code_structure.md
