### 2025-09-16 994bbf7
Update Example_code_structure.md
