### 2025-09-16 c7e1c18
Update Example_code_structure.md
