### 2025-09-16 c12c56d
Update Example_code_structure.md
