### 2025-09-16 b50a3f9
Update Example_code_structure.md
