### 2025-09-16 d917590
Update Example_code_structure.md
