### 2025-09-16 64b4f96
Update model_simulation.md
