### 2025-09-16 de97ca7
Update model_simulation.md
