+++ doc/model_simulation.md
@@ -18,7 +18,7 @@ It returns the simulated output, error, and delay matrix, and is typically used 
 | Name    | Type     | Required | Description |
 |---------|----------|----------|-------------|
 | `model` | struct   | Yes | Identified NARX model structure, e.g. returned by `NonSysID`. |
-| `u`     | vector / matrix | Yes | Input signal(s) used for simulation. The number of columns depends on the number of inputs in the model. |
+| `u`     | vector / matrix | Yes | Input signal(s) used for simulation. The number of columns depends on the number of inputs in the model. If a (N)/AR model is simulated, then set `u` to a column vector of zeros with the same length as `y` (`u=y.*0`) |
 | `y`     | column vector   | Yes | Actual measured output signal (for comparison with simulated output and for one-step-ahead/k-step-ahead predictions). If only model simulation is required, then set `y` to a column vector of zeros with the same length as `u` (`y=u(:,1).*0`)|
 | `KSA_h` | int      | Yes | Horizon for **k-step-ahead prediction** in the simulation. |
 
