### 2025-09-16 bc2a6ca
Update Example_code_structure.md
