+++ Examples/README.md
@@ -1,5 +1,5 @@
 # Examples on using the NonSysID package
-This document provides system identification examples using the NonSysID package. The examples shown here follow the code structure shown in [here](./Example_code_structure.md). This guide shows a step-by-step guide to applying the `NonSysID` package in MATLAB.
+The examples shown here follow the code structure shown in [here](./Example_code_structure.md). This guide shows a step-by-step guide to applying the `NonSysID` package in MATLAB.
 
 
 
