### 2025-09-16 8f10232
Update paper.bib
