+++ paper/paper.bib
@@ -100,7 +100,8 @@ volume = {34},
 number = {1},
 pages = {1-12},
 year = {2010},
-author = {Lennart Ljung}
+author = {Lennart Ljung},
+doi = {10.1016/j.arcontrol.2009.12.001}
 }
 
 @book{landau2006digital,
