### 2025-09-16 b8a577f
Update paper.bib
