+++ paper/paper.bib
@@ -134,7 +134,7 @@ number = {1},
 pages = {26-40},
 year = {1995},
 issn = {0167-2789},
-doi = {https://doi.org/10.1016/0167-2789(95)90053-5},
+doi = {10.1016/0167-2789(95)90053-5},
 author = {Luis Antonio Aguirre and S.A. Billings}
 }
 
@@ -171,7 +171,8 @@ doi = {10.1080/00207170310001639640}
   year = {1989},
   language = {en},
   journal = {Int. J. Control},
-  number = {5}
+  number = {5},
+  doi = {10.1080/00207178908953472}
 }
 
 @misc{billings1987a,
@@ -180,7 +181,8 @@ author = {Billings, S.A. and Korenberg, M.J. and Chen, S.},
 title = {Identification of non-linear output-affine systems using an orthogonal least-squares algorithm},
 year = {1987},
 language = {en},
-address = {Sheffield}
+address = {Sheffield},
+doi = {10.1080/00207728808964057}
 }
 
 @article{korenberg1988a,
@@ -192,7 +194,8 @@ year = {1988},
 pages = {193–210,},
 language = {pt},
 journal = {Int. J. Control},
-number = {1}
+number = {1},
+doi = {10.1080/00207178808906169}
 }
 
 @article{UNBEHAUEN1997,
@@ -204,7 +207,7 @@ pages = {973-999},
 year = {1997},
 note = {IFAC Symposium on System Identification (SYSID'97), Kitakyushu, Fukuoka, Japan, 8-11 July 1997},
 issn = {1474-6670},
-doi = {https://doi.org/10.1016/S1474-6670(17)42970-3},
+doi = {10.1016/S1474-6670(17)42970-3},
 author = {H. Unbehauen and P. Rao}
 }
 
@@ -217,7 +220,7 @@ pages = {427-452},
 year = {2024},
 series = {Woodhead Publishing Series in Composites Science and Engineering},
 isbn = {978-0-443-15476-8},
-doi = {https://doi.org/10.1016/B978-0-443-15476-8.00004-6},
+doi = {10.1016/B978-0-443-15476-8.00004-6},
 author = {Kangwei Wang and Jie Zhang and Anthony J. Croxford and Yong Yang}
 }
 
@@ -241,7 +244,7 @@ volume = {220},
 pages = {106835},
 year = {2022},
 issn = {0169-2607},
-doi = {https://doi.org/10.1016/j.cmpb.2022.106835},
+doi = {10.1016/j.cmpb.2022.106835},
 author = {Nurhidayah Mohd Zainol and Nor Salwa Damanhuri and Nor Azlan Othman and Yeong Shiong Chiew and Mohd Basri Mat Nor and Zuraida Muhammad and J. Geoffrey Chase}
 }
 
@@ -252,7 +255,7 @@ volume = {361},
 pages = {144-152},
 year = {2017},
 issn = {0378-7753},
-doi = {https://doi.org/10.1016/j.jpowsour.2017.06.068},
+doi = {10.1016/j.jpowsour.2017.06.068},
 author = {D. Ritzberger and S. Jakubek}
 }
 
@@ -276,7 +279,7 @@ volume = {324},
 pages = {377-389},
 year = {2016},
 issn = {0306-4522},
-doi = {https://doi.org/10.1016/j.neuroscience.2016.03.028},
+doi = {10.1016/j.neuroscience.2016.03.028},
[…]
