### 2025-09-16 #8028 @raj-gun
Hi @boisgera, @JAuriol81 and @brivadis,

Hope you all are well. Thank you again for taking the time to review the repository and the code and for the valuable feedback.

I have reduced the number of words to the limit required. There is a new folder, '[supplementary_information](https://github.com/raj-gun/NonSysID/tree/main/supplementary_information)' included in the repository. The 'supplementary_information' folder contains the theoretical background and other details removed from the paper. The paper is ready to be reviewed.

A [documentation](https://github.com/raj-gun/NonSysID/tree/main/doc) folder was also added to the repository. This was added to address some of the issues raised.

I have addressed all the issues raised and have closed them.

Thank you again for being flexible with me. Sorry for the inconvenience caused due to the delays in addressing everything that was raised.

Thank you
Kind regards
Raj (@raj-gun)
