### 2025-09-19 #8028 @boisgera
Hi @JAuriol81, @brivadis,

Could you update your checklist when you get the chance and assess if there are remaining/extra issues that need to be considered?  Thanks in advance! 🙏
