### 2025-09-22 #8028 @boisgera
@editoriatbot generate pdf
